refactor(api): extract handlers from mod.rs into dedicated modules
Phase 3 API extraction: moved ~3000 lines of handler implementations
from mod.rs into focused module files:

- system.rs: health checks, version, maintenance, handshake, metrics/
  security/request-ID middleware (345 lines)
- users.rs: profile CRUD, GDPR export/delete, password change,
  preferences, user stats (757 lines)
- admin_handlers.rs: user/booking management, stats, reports, heatmap,
  audit log, settings (auto-release, email, privacy, reset) (1412 lines)
- lots_ext.rs: lot QR code, admin dashboard charts (267 lines)
- misc.rs: legal/Impressum, public occupancy, public display (384 lines)

mod.rs now contains only router setup, middleware wiring, shared types,
and crypto utilities (~1500 lines, down from ~4500).

All 481 backend tests pass, zero clippy warnings.
