+++ parkhub-server/src/api/lots_ext.rs
@@ -1,0 +1,266 @@
+//! Extended lot handlers: QR code generation, admin dashboard charts.
+//!
+//! Extracted from mod.rs — Phase 3 API extraction.
+
+use axum::{
+    extract::{Path, State},
+    http::StatusCode,
+    Extension, Json,
+};
+use chrono::{TimeDelta, Timelike, Utc};
+use serde::Serialize;
+use std::sync::Arc;
+use tokio::sync::RwLock;
+use uuid::Uuid;
+
+use parkhub_common::ApiResponse;
+
+use crate::AppState;
+
+use super::{check_admin, read_admin_setting, AuthUser};
+
+type SharedState = Arc<RwLock<AppState>>;
+
+// ═══════════════════════════════════════════════════════════════════════════════
+// QR CODE GENERATION (EXTERNAL SERVICE)
+// ═══════════════════════════════════════════════════════════════════════════════
+
+/// QR code response with URLs for generating a QR code externally.
+#[derive(Debug, Serialize)]
+pub struct LotQrResponse {
+    /// URL to an external QR code image service.
+    /// NOTE: Uses api.qrserver.com. For privacy-sensitive deployments,
+    /// operators should generate QR codes locally.
+    qr_url: String,
+    /// The lot's booking page URL that the QR code encodes.
+    lot_url: String,
+}
+
+/// `GET /api/v1/lots/{id}/qr` — generate a QR code URL for a parking lot.
+///
+/// Returns a URL pointing to an external QR API (api.qrserver.com) that renders
+/// a QR code linking to the lot's booking page.
+#[utoipa::path(get, path = "/api/v1/lots/{id}/qr", tag = "Lots",
+    summary = "Generate QR code URL for a lot",
+    description = "Returns a URL to an external QR service encoding the lot's booking page.",
+    security(("bearer_auth" = [])), params(("id" = String, Path, description = "Lot UUID")),
+    responses((status = 200, description = "QR URLs"), (status = 404, description = "Lot not found"))
+)]
+pub async fn lot_qr_code(
+    State(state): State<SharedState>,
+    Extension(_auth_user): Extension<AuthUser>,
+    Path(id): Path<String>,
+) -> (StatusCode, Json<ApiResponse<LotQrResponse>>) {
+    let state_guard = state.read().await;
+
+    // Verify lot exists
+    match state_guard.db.get_parking_lot(&id).await {
+        Ok(Some(_)) => {}
+        Ok(None) => {
+            return (
+                StatusCode::NOT_FOUND,
+                Json(ApiResponse::error("NOT_FOUND", "Parking lot not found")),
+            );
+        }
+        Err(e) => {
+            tracing::error!("Database error: {}", e);
+            return (
+                StatusCode::INTERNAL_SERVER_ERROR,
+                Json(ApiResponse::error("SERVER_ERROR", "Internal server error")),
+            );
+        }
+    }
+
+    // Derive base_url from admin setting or fall back to localhost
+    let base_url = read_admin_setting(&state_guard.db, "base_url").await;
+    let base_url = if base_url.is_empty() {
+        format!("https://localhost:{}", state_guard.config.port)
+    } else {
+        base_url
+    };
+
+    let lot_url = format!("{base_url}/book?lot={id}");
+    let encoded = url::form_urlencoded::Serializer::new(String::new())
+        .append_pair("data", &lot_url)
+        .append_pair("size", "256x256")
+        .finish();
+    let qr_url = format!("https://api.qrserver.com/v1/create-qr-code/?{encoded}");
+
+    (
+        StatusCode::OK,
+        Json(ApiResponse::success(LotQrResponse { qr_url, lot_url })),
+    )
+}
+
+// ═══════════════════════════════════════════════════════════════════════════════
+// DASHBOARD CHARTS (ADMIN)
+// ═══════════════════════════════════════════════════════════════════════════════
+
+#[derive(Debug, Serialize)]
+struct BookingsByDay {
+    date: String,
+    count: usize,
+}
+
+#[derive(Debug, Serialize)]
+struct BookingsByLot {
+    lot_name: String,
+    count: usize,
+}
+
+#[derive(Debug, Serialize)]
+struct OccupancyByHour {
+    hour: u32,
+    avg_occupancy: f64,
+}
+
+#[derive(Debug, Serialize)]
+struct TopUser {
+    username: String,
+    booking_count: usize,
+}
+
+#[derive(Debug, Serialize)]
+pub struct DashboardCharts {
+    bookings_by_day: Vec<BookingsByDay>,
+    bookings_by_lot: Vec<BookingsByLot>,
+    occupancy_by_hour: Vec<OccupancyByHour>,
+    top_users: Vec<TopUser>,
+}
+
+/// `GET /api/v1/admin/dashboard/charts` — aggregated chart data for the admin
+/// dashboard.  Returns bookings-by-day (last 30 days), bookings-by-lot,
+/// average occupancy by hour-of-day, and top-10 users by booking count.
+#[utoipa::path(get, path = "/api/v1/admin/dashboard/charts", tag = "Admin",
+    summary = "Admin dashboard chart data",
+    description = "Returns aggregated chart data for the admin dashboard.",
+    security(("bearer_auth" = [])),
+    responses((status = 200, description = "Chart data"), (status = 403, description = "Forbidden"))
+)]
+pub async fn admin_dashboard_charts(
+    State(state): State<SharedState>,
+    Extension(auth_user): Extension<AuthUser>,
+) -> (StatusCode, Json<ApiResponse<DashboardCharts>>) {
+    let state_guard = state.read().await;
+    if let Err((status, msg)) = check_admin(&state_guard, &auth_user).await {
+        return (status, Json(ApiResponse::error("FORBIDDEN", msg)));
+    }
+
+    let bookings = state_guard.db.list_bookings().await.unwrap_or_default();
+    let lots = state_guard.db.list_parking_lots().await.unwrap_or_default();
+    let users = state_guard.db.list_users().await.unwrap_or_default();
+    let now = Utc::now();
+    let cutoff = now - TimeDelta::days(30);
+
+    // ── bookings_by_day (last 30 days) ──────────────────────────────────────
+    let mut by_day: std::collections::BTreeMap<String, usize> = std::collections::BTreeMap::new();
+    // Pre-fill all 30 days so the chart has continuous x-axis
+    for d in 0..30 {
+        let date = (now - TimeDelta::days(d)).format("%Y-%m-%d").to_string();
+        by_day.entry(date).or_insert(0);
+    }
+    for b in &bookings {
+        if b.created_at >= cutoff {
+            let date = b.created_at.format("%Y-%m-%d").to_string();
+            *by_day.entry(date).or_insert(0) += 1;
+        }
+    }
+    let bookings_by_day: Vec<BookingsByDay> = by_day
+        .into_iter()
+        .map(|(date, count)| BookingsByDay { date, count })
+        .collect();
+
+    // ── bookings_by_lot ─────────────────────────────────────────────────────
+    let lot_name_map: std::collections::HashMap<Uuid, String> =
+        lots.iter().map(|l| (l.id, l.name.clone())).collect();
+    let mut by_lot: std::collections::HashMap<String, usize> = std::collections::HashMap::new();
+    for b in &bookings {
+        let name = lot_name_map
+            .get(&b.lot_id)
+            .cloned()
+            .unwrap_or_else(|| b.lot_id.to_string());
+        *by_lot.entry(name).or_insert(0) += 1;
+    }
+    let mut bookings_by_lot: Vec<BookingsByLot> = by_lot
+        .into_iter()
+        .map(|(lot_name, count)| BookingsByLot { lot_name, count })
+        .collect();
+    bookings_by_lot.sort_by(|a, b| b.count.cmp(&a.count));
+
+    // ── occupancy_by_hour (average across all lots) ─────────────────────────
+    // For each hour of the day, count how many bookings are active during that
+    // hour within the last 30 days, then divide by number of days with data.
+    let total_slots: i32 = lots.iter().map(|l| l.total_slots).sum();
+    let mut hour_totals = [0usize; 24];
+    let mut hour_days = [0usize; 24];
+
+    // Count distinct days per hour that had at least one booking
+    let mut hour_day_set: [std::collections::HashSet<String>; 24] =
+        std::array::from_fn(|_| std::collections::HashSet::new());
+
+    for b in &bookings {
+        if b.start_time >= cutoff || b.end_time >= cutoff {
+            // Walk through each hour the booking spans
+            let mut t = b.start_time;
+            while t < b.end_time && t < now {
+                let h = t.hour() as usize;
+                if h < 24 {
+                    hour_totals[h] += 1;
+                    hour_day_set[h].insert(t.format("%Y-%m-%d").to_string());
+                }
+                t += TimeDelta::hours(1);
+            }
+        }
+    }
+
+    for (h, day_set) in hour_day_set.iter().enumerate() {
+        hour_days[h] = day_set.len().max(1);
+    }
+
+    let occupancy_by_hour: Vec<OccupancyByHour> = (0..24)
+        .map(|h| {
+            #[allow(clippy::cast_precision_loss)]
+            let avg_count = hour_totals[h] as f64 / hour_days[h] as f64;
+            let avg_occ = if total_slots > 0 {
+                (avg_count / f64::from(total_slots)).min(1.0)
+            } else {
+                0.0
+            };
+            OccupancyByHour {
+                hour: u32::try_from(h).unwrap_or(0),
+                avg_occupancy: (avg_occ * 100.0).round() / 100.0,
+            }
+        })
+        .collect();
+
+    // ── top_users (top 10 by booking count) ─────────────────────────────────
+    let user_name_map: std::collections::HashMap<Uuid, String> =
+        users.iter().map(|u| (u.id, u.username.clone())).collect();
+    let mut by_user: std::collections::HashMap<String, usize> = std::collections::HashMap::new();
+    for b in &bookings {
+        let name = user_name_map
+            .get(&b.user_id)
+            .cloned()
+            .unwrap_or_else(|| b.user_id.to_string());
+        *by_user.entry(name).or_insert(0) += 1;
+    }
+    let mut top_users: Vec<TopUser> = by_user
+        .into_iter()
+        .map(|(username, booking_count)| TopUser {
+            username,
+            booking_count,
+        })
+        .collect();
+    top_users.sort_by(|a, b| b.booking_count.cmp(&a.booking_count));
+    top_users.truncate(10);
+
+    (
+        StatusCode::OK,
+        Json(ApiResponse::success(DashboardCharts {
+            bookings_by_day,
+            bookings_by_lot,
+            occupancy_by_hour,
+            top_users,
+        })),
+    )
+}
