+++ parkhub-server/src/api/misc.rs
@@ -1,0 +1,383 @@
+//! Miscellaneous public handlers: legal (Impressum), public occupancy, display.
+//!
+//! Extracted from mod.rs — Phase 3 API extraction.
+
+use axum::{
+    extract::State,
+    http::{header, StatusCode},
+    Extension, Json,
+};
+use chrono::Utc;
+use serde::{Deserialize, Serialize};
+use std::fmt::Write as _;
+use std::sync::Arc;
+use tokio::sync::RwLock;
+
+use parkhub_common::{ApiResponse, UserRole};
+
+use crate::AppState;
+
+use super::AuthUser;
+
+type SharedState = Arc<RwLock<AppState>>;
+
+// ═══════════════════════════════════════════════════════════════════════════════
+// LEGAL / IMPRESSUM (DDG § 5)
+// ═══════════════════════════════════════════════════════════════════════════════
+
+/// DDG § 5 Impressum fields stored as settings keys with "impressum_" prefix
+#[derive(Debug, Serialize, Deserialize, Default)]
+#[allow(dead_code)]
+pub struct ImpressumData {
+    pub provider_name: String,
+    pub provider_legal_form: String,
+    pub street: String,
+    pub zip_city: String,
+    pub country: String,
+    pub email: String,
+    pub phone: String,
+    pub register_court: String,
+    pub register_number: String,
+    pub vat_id: String,
+    pub responsible_person: String,
+    pub custom_text: String,
+}
+
+const IMPRESSUM_FIELDS: &[&str] = &[
+    "provider_name",
+    "provider_legal_form",
+    "street",
+    "zip_city",
+    "country",
+    "email",
+    "phone",
+    "register_court",
+    "register_number",
+    "vat_id",
+    "responsible_person",
+    "custom_text",
+];
+
+/// Public Impressum endpoint — no auth required (DDG § 5)
+#[utoipa::path(get, path = "/api/v1/legal/impressum", tag = "Public",
+    summary = "Get Impressum (public)", description = "Returns DDG paragraph 5 Impressum data. No auth required.",
+    responses((status = 200, description = "Impressum fields"))
+)]
+pub async fn get_impressum(State(state): State<SharedState>) -> Json<serde_json::Value> {
+    let mut data = serde_json::json!({});
+    {
+        let state = state.read().await;
+        for field in IMPRESSUM_FIELDS {
+            let key = format!("impressum_{field}");
+            let value = state
+                .db
+                .get_setting(&key)
+                .await
+                .unwrap_or(None)
+                .unwrap_or_default();
+            data[field] = serde_json::Value::String(value);
+        }
+    }
+
+    Json(data)
+}
+
+/// Admin: read Impressum settings (admin-only, protected).
+///
+/// Although the public endpoint exposes the same data, this route is kept
+/// separate so admins can fetch the current values before editing them via PUT.
+/// It is deliberately restricted to Admin/SuperAdmin.
+#[utoipa::path(get, path = "/api/v1/admin/impressum", tag = "Admin",
+    summary = "Get Impressum settings (admin)", description = "Returns current Impressum fields for editing. Admin only.",
+    security(("bearer_auth" = [])),
+    responses((status = 200, description = "Impressum fields"), (status = 403, description = "Forbidden"))
+)]
+pub async fn get_impressum_admin(
+    State(state): State<SharedState>,
+    Extension(auth_user): Extension<AuthUser>,
+) -> (StatusCode, Json<serde_json::Value>) {
+    let state_guard = state.read().await;
+
+    // Verify admin role.
+    let Ok(Some(caller)) = state_guard
+        .db
+        .get_user(&auth_user.user_id.to_string())
+        .await
+    else {
+        return (
+            StatusCode::FORBIDDEN,
+            Json(serde_json::json!({"error": "FORBIDDEN", "message": "Admin access required"})),
+        );
+    };
+
+    if caller.role != UserRole::Admin && caller.role != UserRole::SuperAdmin {
+        return (
+            StatusCode::FORBIDDEN,
+            Json(serde_json::json!({"error": "FORBIDDEN", "message": "Admin access required"})),
+        );
+    }
+
+    let mut data = serde_json::json!({});
+    for field in IMPRESSUM_FIELDS {
+        let key = format!("impressum_{field}");
+        let value = state_guard
+            .db
+            .get_setting(&key)
+            .await
+            .unwrap_or(None)
+            .unwrap_or_default();
+        data[field] = serde_json::Value::String(value);
+    }
+
+    (StatusCode::OK, Json(data))
+}
+
+/// Admin: update Impressum settings
+#[utoipa::path(put, path = "/api/v1/admin/impressum", tag = "Admin",
+    summary = "Update Impressum (admin)", description = "Saves DDG paragraph 5 Impressum fields. Admin only.",
+    security(("bearer_auth" = [])),
+    responses((status = 200, description = "Saved"), (status = 403, description = "Forbidden"))
+)]
+pub async fn update_impressum(
+    State(state): State<SharedState>,
+    Extension(auth_user): Extension<AuthUser>,
+    Json(payload): Json<serde_json::Value>,
+) -> (StatusCode, Json<ApiResponse<()>>) {
+    // Verify admin role
+    let user_id_str = auth_user.user_id.to_string();
+    let state_guard = state.read().await;
+    let Ok(Some(user)) = state_guard.db.get_user(&user_id_str).await else {
+        return (
+            StatusCode::FORBIDDEN,
+            Json(ApiResponse::error("FORBIDDEN", "Admin required")),
+        );
+    };
+    drop(state_guard);
+
+    if user.role != UserRole::Admin && user.role != UserRole::SuperAdmin {
+        return (
+            StatusCode::FORBIDDEN,
+            Json(ApiResponse::error("FORBIDDEN", "Admin required")),
+        );
+    }
+
+    let state_guard = state.read().await;
+    for field in IMPRESSUM_FIELDS {
+        if let Some(serde_json::Value::String(value)) = payload.get(*field) {
+            let key = format!("impressum_{field}");
+            if let Err(e) = state_guard.db.set_setting(&key, value).await {
+                tracing::warn!("Failed to save impressum setting {key}: {e}");
+            }
+        }
+    }
+
+    (StatusCode::OK, Json(ApiResponse::success(())))
+}
+
+// ═══════════════════════════════════════════════════════════════════════════════
+// PUBLIC OCCUPANCY
+// ═══════════════════════════════════════════════════════════════════════════════
+
+/// Occupancy info for a single lot
+#[derive(Debug, Serialize)]
+pub struct LotOccupancy {
+    lot_id: String,
+    lot_name: String,
+    total_slots: i32,
+    occupied_slots: i32,
+    available_slots: i32,
+}
+
+/// `GET /api/v1/public/occupancy` — public JSON occupancy data
+#[utoipa::path(get, path = "/api/v1/public/occupancy", tag = "Public",
+    summary = "Public lot occupancy",
+    description = "Returns real-time occupancy. No auth required.",
+    responses((status = 200, description = "Success"))
+)]
+pub async fn public_occupancy(
+    State(state): State<SharedState>,
+) -> (StatusCode, Json<ApiResponse<Vec<LotOccupancy>>>) {
+    let state_guard = state.read().await;
+
+    let lots = match state_guard.db.list_parking_lots().await {
+        Ok(l) => l,
+        Err(e) => {
+            tracing::error!("Failed to list lots for occupancy: {}", e);
+            return (
+                StatusCode::INTERNAL_SERVER_ERROR,
+                Json(ApiResponse::error(
+                    "SERVER_ERROR",
+                    "Failed to get occupancy",
+                )),
+            );
+        }
+    };
+
+    // Count active bookings per lot
+    let now = Utc::now();
+    let bookings = state_guard.db.list_bookings().await.unwrap_or_default();
+
+    let mut occupancy = Vec::with_capacity(lots.len());
+    for lot in &lots {
+        let occupied = i32::try_from(
+            bookings
+                .iter()
+                .filter(|b| {
+                    b.lot_id == lot.id
+                        && b.start_time <= now
+                        && b.end_time >= now
+                        && matches!(
+                            b.status,
+                            parkhub_common::BookingStatus::Confirmed
+                                | parkhub_common::BookingStatus::Active
+                        )
+                })
+                .count(),
+        )
+        .unwrap_or(i32::MAX);
+
+        let available = (lot.total_slots - occupied).max(0);
+
+        occupancy.push(LotOccupancy {
+            lot_id: lot.id.to_string(),
+            lot_name: lot.name.clone(),
+            total_slots: lot.total_slots,
+            occupied_slots: occupied,
+            available_slots: available,
+        });
+    }
+
+    (StatusCode::OK, Json(ApiResponse::success(occupancy)))
+}
+
+/// `GET /api/v1/public/display` — simplified HTML for parking displays
+#[utoipa::path(get, path = "/api/v1/public/display", tag = "Public",
+    summary = "Public display HTML",
+    description = "Returns minimal HTML for digital signage.",
+    responses((status = 200, description = "Success"))
+)]
+pub async fn public_display(State(state): State<SharedState>) -> impl axum::response::IntoResponse {
+    let state_guard = state.read().await;
+
+    let lots = state_guard.db.list_parking_lots().await.unwrap_or_default();
+    let now = Utc::now();
+    let bookings = state_guard.db.list_bookings().await.unwrap_or_default();
+
+    let mut html = String::from(
+        r#"<!DOCTYPE html>
+<html lang="en">
+<head>
+<meta charset="UTF-8">
+<meta name="viewport" content="width=device-width, initial-scale=1.0">
+<meta http-equiv="refresh" content="30">
+<title>ParkHub — Parking Availability</title>
+<style>
+  body { font-family: system-ui, sans-serif; margin: 0; padding: 2rem; background: #1a1a2e; color: #eee; }
+  h1 { text-align: center; margin-bottom: 2rem; }
+  .lots { display: flex; flex-wrap: wrap; gap: 1.5rem; justify-content: center; }
+  .lot { background: #16213e; border-radius: 12px; padding: 1.5rem 2rem; min-width: 220px; text-align: center; }
+  .lot-name { font-size: 1.2rem; font-weight: 600; margin-bottom: 0.5rem; }
+  .available { font-size: 3rem; font-weight: 700; }
+  .available.green { color: #4ade80; }
+  .available.yellow { color: #facc15; }
+  .available.red { color: #f87171; }
+  .label { font-size: 0.9rem; color: #94a3b8; }
+</style>
+</head>
+<body>
+<h1>Parking Availability</h1>
+<div class="lots">
+"#,
+    );
+
+    for lot in &lots {
+        let occupied = i32::try_from(
+            bookings
+                .iter()
+                .filter(|b| {
+                    b.lot_id == lot.id
+                        && b.start_time <= now
+                        && b.end_time >= now
+                        && matches!(
+                            b.status,
+                            parkhub_common::BookingStatus::Confirmed
+                                | parkhub_common::BookingStatus::Active
+                        )
+                })
+                .count(),
+        )
+        .unwrap_or(i32::MAX);
+
+        let available = (lot.total_slots - occupied).max(0);
+        let pct = if lot.total_slots > 0 {
+            (f64::from(available) / f64::from(lot.total_slots)) * 100.0
+        } else {
+            0.0
+        };
+        let color_class = if pct > 30.0 {
+            "green"
+        } else if pct > 10.0 {
+            "yellow"
+        } else {
+            "red"
+        };
+
+        {
+            let _ = write!(
+                html,
+                r#"<div class="lot">
+  <div class="lot-name">{}</div>
+  <div class="available {}">{}</div>
+  <div class="label">of {} available</div>
+</div>
+"#,
+                crate::utils::html_escape(&lot.name),
+                color_class,
+                available,
+                lot.total_slots
+            );
+        }
+    }
+
+    html.push_str("</div>\n</body>\n</html>\n");
+
+    (
+        StatusCode::OK,
+        [(header::CONTENT_TYPE, "text/html; charset=utf-8")],
+        html,
+    )
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_impressum_data_default() {
+        let data = ImpressumData::default();
+        assert_eq!(data.provider_name, "");
+        assert_eq!(data.country, "");
+    }
+
+    #[test]
+    fn test_impressum_data_roundtrip() {
+        let data = ImpressumData {
+            provider_name: "ParkCorp GmbH".to_string(),
+            provider_legal_form: "GmbH".to_string(),
+            street: "Musterstr. 1".to_string(),
+            zip_city: "12345 Berlin".to_string(),
+            country: "DE".to_string(),
+            email: "info@parkcorp.de".to_string(),
+            phone: "+49 30 123456".to_string(),
+            register_court: "Amtsgericht Berlin".to_string(),
+            register_number: "HRB 12345".to_string(),
+            vat_id: "DE123456789".to_string(),
+            responsible_person: "Max Mustermann".to_string(),
+            custom_text: "".to_string(),
+        };
+        let json = serde_json::to_string(&data).unwrap();
+        let back: ImpressumData = serde_json::from_str(&json).unwrap();
+        assert_eq!(back.provider_name, "ParkCorp GmbH");
+        assert_eq!(back.vat_id, "DE123456789");
+    }
+}
