+++ parkhub-server/src/api/mod.rs
@@ -4,15 +4,13 @@
 
 use axum::{
     body::Body,
-    extract::{Path, Query, State},
+    extract::State,
     http::{header, HeaderName, HeaderValue, Request, StatusCode},
     middleware::{self, Next},
     response::{IntoResponse, Response},
     routing::{delete, get, post, put},
     Extension, Json, Router,
 };
-use chrono::{Datelike, TimeDelta, Timelike, Utc};
-use std::fmt::Write as _;
 use std::sync::Arc;
 use std::time::Duration;
 use tokio::sync::RwLock;
@@ -22,14 +20,12 @@ use tower_http::cors::CorsLayer;
 use tower_http::limit::RequestBodyLimitLayer;
 use tower_http::request_id::{MakeRequestUuid, PropagateRequestIdLayer, SetRequestIdLayer};
 use tower_http::trace::TraceLayer;
-use tracing::Instrument;
 #[cfg(feature = "full")]
 use utoipa::OpenApi;
 #[cfg(feature = "full")]
 use utoipa_swagger_ui::SwaggerUi;
 use uuid::Uuid;
 
-use crate::audit::{AuditEntry, AuditEventType};
 use crate::demo;
 #[cfg(feature = "mod-email")]
 use crate::email;
@@ -52,11 +48,7 @@ pub const MAX_PHOTO_BYTES: usize = 2 * 1024 * 1024;
 #[cfg(feature = "mod-bookings")]
 pub(super) const VAT_RATE: f64 = 0.19;
 
-use parkhub_common::{
-    ApiResponse, BookingStatus, CreditTransactionType, HandshakeRequest, HandshakeResponse,
-    LoginResponse, ServerStatus, User, UserRole, PROTOCOL_VERSION,
-};
-use serde::{Deserialize, Serialize};
+use parkhub_common::{ApiResponse, LoginResponse, UserRole};
 
 use crate::AppState;
 
@@ -70,6 +62,7 @@ type SharedState = Arc<RwLock<AppState>>;
 pub mod absences;
 pub mod admin;
 pub mod admin_ext;
+pub mod admin_handlers;
 #[cfg(feature = "mod-announcements")]
 pub mod announcements;
 pub mod auth;
@@ -94,6 +87,8 @@ pub mod import;
 #[cfg(feature = "mod-invoices")]
 pub mod invoices;
 pub mod lots;
+pub mod lots_ext;
+pub mod misc;
 #[cfg(feature = "mod-notifications")]
 pub mod notification_channels;
 #[cfg(feature = "mod-notifications")]
@@ -122,6 +117,7 @@ pub mod setup;
 mod social;
 #[cfg(feature = "mod-swap")]
 pub mod swap;
+pub mod system;
 #[cfg(feature = "mod-team")]
 pub mod team;
 #[cfg(feature = "mod-translations")]
@@ -232,6 +228,22 @@ use waitlist::{join_waitlist, leave_waitlist, list_waitlist};
 use webhooks::{create_webhook, delete_webhook, list_webhooks, test_webhook, update_webhook};
 #[cfg(feature = "mod-zones")]
 use zones::{create_zone, delete_zone, list_zones, update_zone};
+
+// Re-exports from extracted modules (Phase 3)
+use admin_handlers::{
+    admin_audit_log, admin_delete_user, admin_get_auto_release, admin_get_email_settings,
+    admin_get_privacy, admin_heatmap, admin_list_bookings, admin_list_users, admin_reports,
+    admin_reset, admin_stats, admin_update_auto_release, admin_update_email_settings,
+    admin_update_privacy, admin_update_user, admin_update_user_role, admin_update_user_status,
+};
+use lots_ext::{admin_dashboard_charts, lot_qr_code};
+use misc::{
+    get_impressum, get_impressum_admin, public_display, public_occupancy, update_impressum,
+};
+use users::{
+    change_password, gdpr_delete_account, gdpr_export_data, get_current_user, get_user,
+    get_user_preferences, update_current_user, update_user_preferences, user_stats,
+};
 
 /// User ID extracted from auth token
 #[derive(Clone, Debug)]
@@ -1214,143 +1226,8 @@ pub fn create_router(state: SharedState) -> (Router, demo::SharedDemoState) {
     (router, demo_state_ret)
 }
 
-/// Middleware that adds security-related response headers to every request.
-///
-/// - `X-Content-Type-Options`: prevents MIME sniffing
-/// - `X-Frame-Options`: prevents clickjacking
-/// - `Content-Security-Policy`: restricts resource origins
-/// - `Referrer-Policy`: limits referrer leakage
-/// - `Permissions-Policy`: disables unneeded browser features
-/// - `Strict-Transport-Security`: enforces HTTPS for 1 year including subdomains
-async fn security_headers_middleware(request: Request<Body>, next: Next) -> Response {
-    let mut response = next.run(request).await;
-    let headers = response.headers_mut();
-
-    // Safety: all header names and values below are valid ASCII strings.
-    headers.insert(
-        HeaderName::from_static("x-content-type-options"),
-        HeaderValue::from_static("nosniff"),
-    );
-    headers.insert(
-        HeaderName::from_static("x-frame-options"),
-        HeaderValue::from_static("DENY"),
-    );
-    headers.insert(
-        HeaderName::from_static("content-security-policy"),
-        HeaderValue::from_static(
-            // 'unsafe-inline' removed from both script-src and style-src (issue #106).
-            // Inline styles should use external stylesheets or CSS-in-JS that injects
-            // <link> elements. If a framework requires inline styles, add specific
-            // hashes via 'unsafe-hashes' rather than blanket 'unsafe-inline'.
-            "default-src 'self'; \
-             script-src 'self'; \
-             style-src 'self'; \
-             img-src 'self' data:; \
-             font-src 'self'; \
-             connect-src 'self' ws: wss:; \
-             frame-ancestors 'none'",
-        ),
-    );
-    headers.insert(
-        HeaderName::from_static("referrer-policy"),
-        HeaderValue::from_static("strict-origin-when-cross-origin"),
-    );
-    headers.insert(
-        HeaderName::from_static("permissions-policy"),
-        HeaderValue::from_static("geolocation=(), camera=(), microphone=()"),
-    );
-    headers.insert(
-        HeaderName::from_static("strict-transport-security"),
-        HeaderValue::from_static("max-age=31536000; includeSubDomains; preload"),
-    );
-
-    response
-}
-
-// ═══════════════════════════════════════════════════════════════════════════════
-// REQUEST ID TRACING MIDDLEWARE
-// ═══════════════════════════════════════════════════════════════════════════════
-
-/// Attaches the `x-request-id` header value (set by `SetRequestIdLayer`) to the
-/// current tracing span so that every log line within this request includes the
-/// request ID.  Also ensures the header is copied to the response.
-async fn request_id_tracing_middleware(request: Request<Body>, next: Next) -> Response {
-    let request_id = request
-        .headers()
-        .get("x-request-id")
-        .and_then(|v| v.to_str().ok())
-        .unwrap_or("-")
-        .to_owned();
-
-    let span = tracing::info_span!("request", request_id = %request_id);
-    let guard = span.enter();
-
-    let mut response = {
-        // Drop the span guard before awaiting so the span covers the full
-        // request lifecycle via the instrument approach.
-        drop(guard);
-        next.run(request).instrument(span).await
-    };
-
-    // Ensure x-request-id is on the response (belt-and-suspenders with PropagateRequestIdLayer)
-    if let Ok(val) = HeaderValue::from_str(&request_id) {
-        response
-            .headers_mut()
-            .entry(HeaderName::from_static("x-request-id"))
-            .or_insert(val);
-    }
-
-    response
-}
-
-// ═══════════════════════════════════════════════════════════════════════════════
-// HTTP METRICS MIDDLEWARE
-// ═══════════════════════════════════════════════════════════════════════════════
-
-/// Records HTTP request metrics (method, path, status, duration) for Prometheus
-/// and emits a structured log line for every request.
-async fn http_metrics_middleware(request: Request<Body>, next: Next) -> Response {
-    let method = request.method().to_string();
-    let path = request.uri().path().to_string();
-    let start = std::time::Instant::now();
-
-    let response = next.run(request).await;
-
-    let status = response.status().as_u16();
-    let duration = start.elapsed();
-
-    // Normalize path to avoid high-cardinality labels (strip UUIDs/IDs)
-    let normalized = normalize_metric_path(&path);
-    metrics::record_http_request(&method, &normalized, status, duration);
-
-    // Structured request log — every request gets one line with key fields
-    tracing::info!(
-        http.method = %method,
-        http.path = %path,
-        http.status = status,
-        http.latency_ms = u64::try_from(duration.as_millis()).unwrap_or(u64::MAX),
-        "request completed"
-    );
-
-    response
-}
-
-/// Collapse dynamic path segments (UUIDs, numeric IDs) into placeholders
-/// to keep Prometheus label cardinality bounded.
-fn normalize_metric_path(path: &str) -> String {
-    path.split('/')
-        .map(|s| {
-            let is_uuid = s.len() == 36 && s.chars().filter(|c| *c == '-').count() == 4;
-            let is_numeric = !s.is_empty() && s.chars().all(|c| c.is_ascii_digit());
-            if is_uuid || is_numeric {
-                ":id"
-            } else {
-                s
-            }
-        })
-        .collect::<Vec<_>>()
-        .join("/")
-}
+// Middleware re-exports from system module
+use system::{http_metrics_middleware, request_id_tracing_middleware, security_headers_middleware};
 
 // ═══════════════════════════════════════════════════════════════════════════════
 // AUTH MIDDLEWARE
@@ -1494,624 +1371,11 @@ async fn auth_middleware(
     Ok(next.run(request).await)
 }
 
-// ═══════════════════════════════════════════════════════════════════════════════
-// HEALTH & DISCOVERY
-// ═══════════════════════════════════════════════════════════════════════════════
-
-#[utoipa::path(
-    get,
-    path = "/health",
-    tag = "Health",
-    summary = "Basic health check",
-    description = "Returns 200 OK when the server is running.",
-    responses((status = 200, description = "Healthy"))
-)]
-#[tracing::instrument]
-pub async fn health_check() -> &'static str {
-    "OK"
-}
-
-/// Kubernetes liveness probe - just checks if the service is running
-#[utoipa::path(
-    get,
-    path = "/health/live",
-    tag = "Health",
-    summary = "Kubernetes liveness probe",
-    description = "Returns 200 if the process is alive.",
-    responses((status = 200, description = "Alive"))
-)]
-#[tracing::instrument]
-pub async fn liveness_check() -> StatusCode {
-    StatusCode::OK
-}
-
-/// Kubernetes readiness probe - checks if the service can handle traffic.
-///
-/// Returns only a boolean `ready` field. Internal error details are logged
-/// server-side but never exposed in the response body.
-#[utoipa::path(
-    get,
-    path = "/health/ready",
-    tag = "Health",
-    summary = "Kubernetes readiness probe",
-    description = "Returns 200 when the service can accept traffic.",
-    responses(
-        (status = 200, description = "Ready"),
-        (status = 503, description = "Not ready")
-    )
-)]
-#[tracing::instrument(skip(state))]
-pub async fn readiness_check(State(state): State<SharedState>) -> impl IntoResponse {
-    let state = state.read().await;
-    match state.db.stats().await {
-        Ok(_) => (StatusCode::OK, Json(serde_json::json!({"ready": true}))),
-        Err(e) => {
-            tracing::error!(error = %e, "Readiness check failed — database unavailable");
-            (
-                StatusCode::SERVICE_UNAVAILABLE,
-                Json(serde_json::json!({"ready": false})),
-            )
-        }
-    }
-}
-
-/// `GET /api/v1/system/version` — server version information
-pub async fn system_version() -> Json<serde_json::Value> {
-    Json(serde_json::json!({
-        "version": env!("CARGO_PKG_VERSION"),
-        "name": env!("CARGO_PKG_NAME"),
-    }))
-}
-
-/// `GET /api/v1/system/maintenance` — maintenance mode status
-pub async fn system_maintenance(State(state): State<SharedState>) -> Json<serde_json::Value> {
-    let state = state.read().await;
-    let maintenance = match state.db.get_setting("maintenance_mode").await {
-        Ok(Some(v)) => v == "true",
-        _ => false,
-    };
-    Json(serde_json::json!({
-        "maintenance_mode": maintenance,
-        "message": if maintenance { "System is under maintenance" } else { "" }
-    }))
-}
-
-#[utoipa::path(
-    post,
-    path = "/handshake",
-    tag = "Health",
-    summary = "Protocol handshake",
-    description = "Verifies protocol version compatibility between client and server.",
-    responses((status = 200, description = "Handshake result"))
-)]
-pub async fn handshake(
-    State(state): State<SharedState>,
-    Json(request): Json<HandshakeRequest>,
-) -> Json<ApiResponse<HandshakeResponse>> {
-    let state = state.read().await;
-
-    // Check protocol version compatibility
-    if request.protocol_version != PROTOCOL_VERSION {
-        return Json(ApiResponse::error(
-            "PROTOCOL_MISMATCH",
-            format!(
-                "Protocol version mismatch: server={}, client={}",
-                PROTOCOL_VERSION, request.protocol_version
-            ),
-        ));
-    }
-
-    Json(ApiResponse::success(HandshakeResponse {
-        server_name: state.config.server_name.clone(),
-        server_version: env!("CARGO_PKG_VERSION").to_string(),
-        protocol_version: PROTOCOL_VERSION.to_string(),
-        requires_auth: true,
-        certificate_fingerprint: String::new(),
-    }))
-}
-
-#[utoipa::path(
-    get,
-    path = "/status",
-    tag = "Health",
-    summary = "Server status overview",
-    description = "Returns aggregate server statistics.",
-    responses((status = 200, description = "Server status"))
-)]
-pub async fn server_status(State(state): State<SharedState>) -> Json<ApiResponse<ServerStatus>> {
-    let db_stats = {
-        let state = state.read().await;
-        state.db.stats().await.unwrap_or(crate::db::DatabaseStats {
-            users: 0,
-            bookings: 0,
-            parking_lots: 0,
-            slots: 0,
-            sessions: 0,
-            vehicles: 0,
-        })
-    };
-
-    Json(ApiResponse::success(ServerStatus {
-        uptime_seconds: 0,
-        connected_clients: 0,
-        total_users: u32::try_from(db_stats.users).unwrap_or(u32::MAX),
-        total_bookings: u32::try_from(db_stats.bookings).unwrap_or(u32::MAX),
-        database_size_bytes: 0,
-    }))
-}
-
-// ═══════════════════════════════════════════════════════════════════════════════
-// USERS
-// ═══════════════════════════════════════════════════════════════════════════════
-
-#[utoipa::path(
-    get,
-    path = "/api/v1/users/me",
-    tag = "Users",
-    summary = "Get current user profile",
-    description = "Returns the authenticated user's profile.",
-    security(("bearer_auth" = [])),
-    responses(
-        (status = 200, description = "User profile"),
-        (status = 404, description = "User not found")
-    )
-)]
-#[tracing::instrument(skip(state), fields(user_id = %auth_user.user_id))]
-pub async fn get_current_user(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-) -> (StatusCode, Json<ApiResponse<User>>) {
-    let state = state.read().await;
-
-    match state.db.get_user(&auth_user.user_id.to_string()).await {
-        Ok(Some(mut user)) => {
-            user.password_hash = String::new();
-            (StatusCode::OK, Json(ApiResponse::success(user)))
-        }
-        Ok(None) => (
-            StatusCode::NOT_FOUND,
-            Json(ApiResponse::error("NOT_FOUND", "User not found")),
-        ),
-        Err(e) => {
-            tracing::error!(error = %e, "Failed to fetch current user");
-            (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ApiResponse::error("SERVER_ERROR", "Internal server error")),
-            )
-        }
-    }
-}
-
-/// Request body for updating the current user's profile
-#[derive(Debug, Deserialize, utoipa::ToSchema)]
-pub struct UpdateCurrentUserRequest {
-    name: Option<String>,
-    phone: Option<String>,
-    picture: Option<String>,
-}
-
-/// `PUT /api/v1/users/me` — update the authenticated user's own profile.
-///
-/// Allows users to update their display name, phone number, and profile
-/// picture URL. Fields not included in the request body are left unchanged.
-/// Returns the updated user record (without `password_hash`).
-#[utoipa::path(
-    put,
-    path = "/api/v1/users/me",
-    tag = "Users",
-    summary = "Update current user profile",
-    description = "Updates the authenticated user's display name, phone, and/or profile picture.",
-    security(("bearer_auth" = [])),
-    responses(
-        (status = 200, description = "Profile updated"),
-        (status = 400, description = "Invalid input"),
-        (status = 404, description = "User not found")
-    )
-)]
-pub async fn update_current_user(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-    Json(req): Json<UpdateCurrentUserRequest>,
-) -> (StatusCode, Json<ApiResponse<User>>) {
-    let state_guard = state.read().await;
-
-    let mut user = match state_guard
-        .db
-        .get_user(&auth_user.user_id.to_string())
-        .await
-    {
-        Ok(Some(u)) => u,
-        Ok(None) => {
-            return (
-                StatusCode::NOT_FOUND,
-                Json(ApiResponse::error("NOT_FOUND", "User not found")),
-            );
-        }
-        Err(e) => {
-            tracing::error!("Database error fetching user for update: {}", e);
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ApiResponse::error("SERVER_ERROR", "Internal server error")),
-            );
-        }
-    };
-
-    // ── Input length validation (issue #115) ────────────────────────────────
-    if let Some(ref name) = req.name {
-        if name.len() > 100 {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(ApiResponse::error(
-                    "INVALID_INPUT",
-                    "Name must be at most 100 characters",
-                )),
-            );
-        }
-    }
-    if let Some(ref phone) = req.phone {
-        if phone.len() > 20 {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(ApiResponse::error(
-                    "INVALID_INPUT",
-                    "Phone number must be at most 20 characters",
-                )),
-            );
-        }
-    }
-
-    // Apply only the fields provided in the request
-    if let Some(name) = req.name {
-        user.name = name;
-    }
-    if let Some(phone) = req.phone {
-        user.phone = Some(phone);
-    }
-    if let Some(picture) = req.picture {
-        // Validate picture URL: must be empty, or a well-formed http(s) URL
-        // capped at 2048 characters to prevent abuse.
-        if !picture.is_empty() {
-            if picture.len() > 2048 {
-                return (
-                    StatusCode::BAD_REQUEST,
-                    Json(ApiResponse::error(
-                        "INVALID_INPUT",
-                        "Picture URL must be at most 2048 characters",
-                    )),
-                );
-            }
-            if !picture.starts_with("https://") && !picture.starts_with("http://") {
-                return (
-                    StatusCode::BAD_REQUEST,
-                    Json(ApiResponse::error(
-                        "INVALID_INPUT",
-                        "Picture must be a valid HTTP or HTTPS URL",
-                    )),
-                );
-            }
-        }
-        user.picture = if picture.is_empty() {
-            None
-        } else {
-            Some(picture)
-        };
-    }
-    user.updated_at = Utc::now();
-
-    if let Err(e) = state_guard.db.save_user(&user).await {
-        tracing::error!("Failed to save user profile update: {}", e);
-        return (
-            StatusCode::INTERNAL_SERVER_ERROR,
-            Json(ApiResponse::error(
-                "SERVER_ERROR",
-                "Failed to update profile",
-            )),
-        );
-    }
-
-    AuditEntry::new(AuditEventType::UserUpdated)
-        .user(user.id, &user.username)
-        .log();
-
-    user.password_hash = String::new();
-    (StatusCode::OK, Json(ApiResponse::success(user)))
-}
-
-/// Retrieve a user by ID.
-///
-/// Restricted to Admin and `SuperAdmin` roles. Regular users must use
-/// `GET /api/v1/users/me` to access their own profile.
-#[utoipa::path(get, path = "/api/v1/users/{id}", tag = "Admin",
-    summary = "Get user by ID (admin)",
-    description = "Retrieves any user's profile. Admin/SuperAdmin only.",
-    security(("bearer_auth" = [])),
-    params(("id" = String, Path, description = "User UUID")),
-    responses((status = 200, description = "User found"), (status = 403, description = "Forbidden"), (status = 404, description = "Not found"))
-)]
-pub async fn get_user(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-    Path(id): Path<String>,
-) -> (StatusCode, Json<ApiResponse<User>>) {
-    let state = state.read().await;
-
-    // Verify caller is an admin before exposing arbitrary user records.
-    let Ok(Some(caller)) = state.db.get_user(&auth_user.user_id.to_string()).await else {
-        return (
-            StatusCode::FORBIDDEN,
-            Json(ApiResponse::error("FORBIDDEN", "Access denied")),
-        );
-    };
-
-    if caller.role != UserRole::Admin && caller.role != UserRole::SuperAdmin {
-        return (
-            StatusCode::FORBIDDEN,
-            Json(ApiResponse::error("FORBIDDEN", "Admin access required")),
-        );
-    }
-
-    match state.db.get_user(&id).await {
-        Ok(Some(mut user)) => {
-            user.password_hash = String::new();
-            (StatusCode::OK, Json(ApiResponse::success(user)))
-        }
-        Ok(None) => (
-            StatusCode::NOT_FOUND,
-            Json(ApiResponse::error("NOT_FOUND", "User not found")),
-        ),
-        Err(e) => {
-            tracing::error!("Database error: {}", e);
-            (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ApiResponse::error("SERVER_ERROR", "Internal server error")),
-            )
-        }
-    }
-}
-
-// ═══════════════════════════════════════════════════════════════════════════════
-// QR CODE GENERATION (EXTERNAL SERVICE)
-// ═══════════════════════════════════════════════════════════════════════════════
-
-/// QR code response with URLs for generating a QR code externally.
-#[derive(Debug, Serialize)]
-pub struct LotQrResponse {
-    /// URL to an external QR code image service.
-    /// NOTE: Uses api.qrserver.com. For privacy-sensitive deployments,
-    /// operators should generate QR codes locally.
-    qr_url: String,
-    /// The lot's booking page URL that the QR code encodes.
-    lot_url: String,
-}
-
-/// `GET /api/v1/lots/{id}/qr` — generate a QR code URL for a parking lot.
-///
-/// Returns a URL pointing to an external QR API (api.qrserver.com) that renders
-/// a QR code linking to the lot's booking page.
-#[utoipa::path(get, path = "/api/v1/lots/{id}/qr", tag = "Lots",
-    summary = "Generate QR code URL for a lot",
-    description = "Returns a URL to an external QR service encoding the lot's booking page.",
-    security(("bearer_auth" = [])), params(("id" = String, Path, description = "Lot UUID")),
-    responses((status = 200, description = "QR URLs"), (status = 404, description = "Lot not found"))
-)]
-pub async fn lot_qr_code(
-    State(state): State<SharedState>,
-    Extension(_auth_user): Extension<AuthUser>,
-    Path(id): Path<String>,
-) -> (StatusCode, Json<ApiResponse<LotQrResponse>>) {
-    let state_guard = state.read().await;
-
-    // Verify lot exists
-    match state_guard.db.get_parking_lot(&id).await {
-        Ok(Some(_)) => {}
-        Ok(None) => {
-            return (
-                StatusCode::NOT_FOUND,
-                Json(ApiResponse::error("NOT_FOUND", "Parking lot not found")),
-            );
-        }
-        Err(e) => {
-            tracing::error!("Database error: {}", e);
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ApiResponse::error("SERVER_ERROR", "Internal server error")),
-            );
-        }
-    }
-
-    // Derive base_url from admin setting or fall back to localhost
-    let base_url = read_admin_setting(&state_guard.db, "base_url").await;
-    let base_url = if base_url.is_empty() {
-        format!("https://localhost:{}", state_guard.config.port)
-    } else {
-        base_url
-    };
-
-    let lot_url = format!("{base_url}/book?lot={id}");
-    let encoded = url::form_urlencoded::Serializer::new(String::new())
-        .append_pair("data", &lot_url)
-        .append_pair("size", "256x256")
-        .finish();
-    let qr_url = format!("https://api.qrserver.com/v1/create-qr-code/?{encoded}");
-
-    (
-        StatusCode::OK,
-        Json(ApiResponse::success(LotQrResponse { qr_url, lot_url })),
-    )
-}
-
-// ═══════════════════════════════════════════════════════════════════════════════
-// DASHBOARD CHARTS (ADMIN)
-// ═══════════════════════════════════════════════════════════════════════════════
-
-#[derive(Debug, Serialize)]
-struct BookingsByDay {
-    date: String,
-    count: usize,
-}
-
-#[derive(Debug, Serialize)]
-struct BookingsByLot {
-    lot_name: String,
-    count: usize,
-}
-
-#[derive(Debug, Serialize)]
-struct OccupancyByHour {
-    hour: u32,
-    avg_occupancy: f64,
-}
-
-#[derive(Debug, Serialize)]
-struct TopUser {
-    username: String,
-    booking_count: usize,
-}
-
-#[derive(Debug, Serialize)]
-pub struct DashboardCharts {
-    bookings_by_day: Vec<BookingsByDay>,
-    bookings_by_lot: Vec<BookingsByLot>,
-    occupancy_by_hour: Vec<OccupancyByHour>,
-    top_users: Vec<TopUser>,
-}
-
-/// `GET /api/v1/admin/dashboard/charts` — aggregated chart data for the admin
-/// dashboard.  Returns bookings-by-day (last 30 days), bookings-by-lot,
-/// average occupancy by hour-of-day, and top-10 users by booking count.
-#[utoipa::path(get, path = "/api/v1/admin/dashboard/charts", tag = "Admin",
-    summary = "Admin dashboard chart data",
-    description = "Returns aggregated chart data for the admin dashboard.",
-    security(("bearer_auth" = [])),
-    responses((status = 200, description = "Chart data"), (status = 403, description = "Forbidden"))
-)]
-pub async fn admin_dashboard_charts(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-) -> (StatusCode, Json<ApiResponse<DashboardCharts>>) {
-    let state_guard = state.read().await;
-    if let Err((status, msg)) = check_admin(&state_guard, &auth_user).await {
-        return (status, Json(ApiResponse::error("FORBIDDEN", msg)));
-    }
-
-    let bookings = state_guard.db.list_bookings().await.unwrap_or_default();
-    let lots = state_guard.db.list_parking_lots().await.unwrap_or_default();
-    let users = state_guard.db.list_users().await.unwrap_or_default();
-    let now = Utc::now();
-    let cutoff = now - TimeDelta::days(30);
-
-    // ── bookings_by_day (last 30 days) ──────────────────────────────────────
-    let mut by_day: std::collections::BTreeMap<String, usize> = std::collections::BTreeMap::new();
-    // Pre-fill all 30 days so the chart has continuous x-axis
-    for d in 0..30 {
-        let date = (now - TimeDelta::days(d)).format("%Y-%m-%d").to_string();
-        by_day.entry(date).or_insert(0);
-    }
-    for b in &bookings {
-        if b.created_at >= cutoff {
-            let date = b.created_at.format("%Y-%m-%d").to_string();
-            *by_day.entry(date).or_insert(0) += 1;
-        }
-    }
-    let bookings_by_day: Vec<BookingsByDay> = by_day
-        .into_iter()
-        .map(|(date, count)| BookingsByDay { date, count })
-        .collect();
-
-    // ── bookings_by_lot ─────────────────────────────────────────────────────
-    let lot_name_map: std::collections::HashMap<Uuid, String> =
-        lots.iter().map(|l| (l.id, l.name.clone())).collect();
-    let mut by_lot: std::collections::HashMap<String, usize> = std::collections::HashMap::new();
-    for b in &bookings {
-        let name = lot_name_map
-            .get(&b.lot_id)
-            .cloned()
-            .unwrap_or_else(|| b.lot_id.to_string());
-        *by_lot.entry(name).or_insert(0) += 1;
-    }
-    let mut bookings_by_lot: Vec<BookingsByLot> = by_lot
-        .into_iter()
-        .map(|(lot_name, count)| BookingsByLot { lot_name, count })
-        .collect();
-    bookings_by_lot.sort_by(|a, b| b.count.cmp(&a.count));
-
-    // ── occupancy_by_hour (average across all lots) ─────────────────────────
-    // For each hour of the day, count how many bookings are active during that
-    // hour within the last 30 days, then divide by number of days with data.
-    let total_slots: i32 = lots.iter().map(|l| l.total_slots).sum();
-    let mut hour_totals = [0usize; 24];
-    let mut hour_days = [0usize; 24];
-
-    // Count distinct days per hour that had at least one booking
-    let mut hour_day_set: [std::collections::HashSet<String>; 24] =
-        std::array::from_fn(|_| std::collections::HashSet::new());
-
-    for b in &bookings {
-        if b.start_time >= cutoff || b.end_time >= cutoff {
-            // Walk through each hour the booking spans
-            let mut t = b.start_time;
-            while t < b.end_time && t < now {
-                let h = t.hour() as usize;
-                if h < 24 {
-                    hour_totals[h] += 1;
-                    hour_day_set[h].insert(t.format("%Y-%m-%d").to_string());
-                }
-                t += TimeDelta::hours(1);
-            }
-        }
-    }
-
-    for (h, day_set) in hour_day_set.iter().enumerate() {
-        hour_days[h] = day_set.len().max(1);
-    }
-
-    let occupancy_by_hour: Vec<OccupancyByHour> = (0..24)
-        .map(|h| {
-            #[allow(clippy::cast_precision_loss)]
-            let avg_count = hour_totals[h] as f64 / hour_days[h] as f64;
-            let avg_occ = if total_slots > 0 {
-                (avg_count / f64::from(total_slots)).min(1.0)
-            } else {
-                0.0
-            };
-            OccupancyByHour {
-                hour: u32::try_from(h).unwrap_or(0),
-                avg_occupancy: (avg_occ * 100.0).round() / 100.0,
-            }
-        })
-        .collect();
-
-    // ── top_users (top 10 by booking count) ─────────────────────────────────
-    let user_name_map: std::collections::HashMap<Uuid, String> =
-        users.iter().map(|u| (u.id, u.username.clone())).collect();
-    let mut by_user: std::collections::HashMap<String, usize> = std::collections::HashMap::new();
-    for b in &bookings {
-        let name = user_name_map
-            .get(&b.user_id)
-            .cloned()
-            .unwrap_or_else(|| b.user_id.to_string());
-        *by_user.entry(name).or_insert(0) += 1;
-    }
-    let mut top_users: Vec<TopUser> = by_user
-        .into_iter()
-        .map(|(username, booking_count)| TopUser {
-            username,
-            booking_count,
-        })
-        .collect();
-    top_users.sort_by(|a, b| b.booking_count.cmp(&a.booking_count));
-    top_users.truncate(10);
-
-    (
-        StatusCode::OK,
-        Json(ApiResponse::success(DashboardCharts {
-            bookings_by_day,
-            bookings_by_lot,
-            occupancy_by_hour,
-            top_users,
-        })),
-    )
-}
+// Health & system handler re-exports from system module
+use system::{
+    handshake, health_check, liveness_check, readiness_check, server_status, system_maintenance,
+    system_version,
+};
 
 // ═══════════════════════════════════════════════════════════════════════════════
 // TOKEN GENERATION
@@ -2229,2289 +1493,5 @@ fn verify_password_sync(password: &str, hash: &str) -> bool {
 }
 
 // ═══════════════════════════════════════════════════════════════════════════════
-// LEGAL / IMPRESSUM (DDG § 5)
-// ═══════════════════════════════════════════════════════════════════════════════
-
-/// DDG § 5 Impressum fields stored as settings keys with "impressum_" prefix
-#[derive(Debug, Serialize, Deserialize, Default)]
-#[allow(dead_code)]
-pub struct ImpressumData {
-    pub provider_name: String,
-    pub provider_legal_form: String,
-    pub street: String,
-    pub zip_city: String,
-    pub country: String,
-    pub email: String,
-    pub phone: String,
-    pub register_court: String,
-    pub register_number: String,
-    pub vat_id: String,
-    pub responsible_person: String,
-    pub custom_text: String,
-}
-
-const IMPRESSUM_FIELDS: &[&str] = &[
-    "provider_name",
-    "provider_legal_form",
-    "street",
-    "zip_city",
-    "country",
-    "email",
-    "phone",
-    "register_court",
-    "register_number",
-    "vat_id",
-    "responsible_person",
-    "custom_text",
-];
-
-/// Public Impressum endpoint — no auth required (DDG § 5)
-#[utoipa::path(get, path = "/api/v1/legal/impressum", tag = "Public",
-    summary = "Get Impressum (public)", description = "Returns DDG paragraph 5 Impressum data. No auth required.",
-    responses((status = 200, description = "Impressum fields"))
-)]
-pub async fn get_impressum(State(state): State<SharedState>) -> Json<serde_json::Value> {
-    let mut data = serde_json::json!({});
-    {
-        let state = state.read().await;
-        for field in IMPRESSUM_FIELDS {
-            let key = format!("impressum_{field}");
-            let value = state
-                .db
-                .get_setting(&key)
-                .await
-                .unwrap_or(None)
-                .unwrap_or_default();
-            data[field] = serde_json::Value::String(value);
-        }
-    }
-
-    Json(data)
-}
-
-/// Admin: read Impressum settings (admin-only, protected).
-///
-/// Although the public endpoint exposes the same data, this route is kept
-/// separate so admins can fetch the current values before editing them via PUT.
-/// It is deliberately restricted to Admin/SuperAdmin.
-#[utoipa::path(get, path = "/api/v1/admin/impressum", tag = "Admin",
-    summary = "Get Impressum settings (admin)", description = "Returns current Impressum fields for editing. Admin only.",
-    security(("bearer_auth" = [])),
-    responses((status = 200, description = "Impressum fields"), (status = 403, description = "Forbidden"))
-)]
-pub async fn get_impressum_admin(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-) -> (StatusCode, Json<serde_json::Value>) {
-    let state_guard = state.read().await;
-
-    // Verify admin role.
-    let Ok(Some(caller)) = state_guard
-        .db
-        .get_user(&auth_user.user_id.to_string())
-        .await
-    else {
-        return (
-            StatusCode::FORBIDDEN,
-            Json(serde_json::json!({"error": "FORBIDDEN", "message": "Admin access required"})),
-        );
-    };
-
-    if caller.role != UserRole::Admin && caller.role != UserRole::SuperAdmin {
-        return (
-            StatusCode::FORBIDDEN,
-            Json(serde_json::json!({"error": "FORBIDDEN", "message": "Admin access required"})),
-        );
-    }
-
-    let mut data = serde_json::json!({});
-    for field in IMPRESSUM_FIELDS {
-        let key = format!("impressum_{field}");
-        let value = state_guard
-            .db
-            .get_setting(&key)
-            .await
-            .unwrap_or(None)
-            .unwrap_or_default();
-        data[field] = serde_json::Value::String(value);
-    }
-
-    (StatusCode::OK, Json(data))
-}
-
-/// Admin: update Impressum settings
-#[utoipa::path(put, path = "/api/v1/admin/impressum", tag = "Admin",
-    summary = "Update Impressum (admin)", description = "Saves DDG paragraph 5 Impressum fields. Admin only.",
-    security(("bearer_auth" = [])),
-    responses((status = 200, description = "Saved"), (status = 403, description = "Forbidden"))
-)]
-pub async fn update_impressum(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-    Json(payload): Json<serde_json::Value>,
-) -> (StatusCode, Json<ApiResponse<()>>) {
-    // Verify admin role
-    let user_id_str = auth_user.user_id.to_string();
-    let state_guard = state.read().await;
-    let Ok(Some(user)) = state_guard.db.get_user(&user_id_str).await else {
-        return (
-            StatusCode::FORBIDDEN,
-            Json(ApiResponse::error("FORBIDDEN", "Admin required")),
-        );
-    };
-    drop(state_guard);
-
-    if user.role != UserRole::Admin && user.role != UserRole::SuperAdmin {
-        return (
-            StatusCode::FORBIDDEN,
-            Json(ApiResponse::error("FORBIDDEN", "Admin required")),
-        );
-    }
-
-    let state_guard = state.read().await;
-    for field in IMPRESSUM_FIELDS {
-        if let Some(serde_json::Value::String(value)) = payload.get(*field) {
-            let key = format!("impressum_{field}");
-            if let Err(e) = state_guard.db.set_setting(&key, value).await {
-                tracing::warn!("Failed to save impressum setting {key}: {e}");
-            }
-        }
-    }
-
-    (StatusCode::OK, Json(ApiResponse::success(())))
-}
-
-// ═══════════════════════════════════════════════════════════════════════════════
-// GDPR — Art. 15 (Data Export) + Art. 17 (Right to Erasure)
-// ═══════════════════════════════════════════════════════════════════════════════
-
-/// GDPR Art. 15 — Export all personal data for the authenticated user
-#[utoipa::path(get, path = "/api/v1/users/me/export", tag = "Users",
-    summary = "GDPR data export (Art. 15)", description = "Exports all personal data as JSON download.",
-    security(("bearer_auth" = [])),
-    responses((status = 200, description = "JSON data export"))
-)]
-pub async fn gdpr_export_data(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-) -> impl IntoResponse {
-    let state = state.read().await;
-    let user_id = auth_user.user_id.to_string();
-
-    let Ok(Some(user)) = state.db.get_user(&user_id).await else {
-        return (
-            StatusCode::NOT_FOUND,
-            [(header::CONTENT_TYPE, "application/json")],
-            serde_json::to_string(&ApiResponse::<()>::error("NOT_FOUND", "User not found"))
-                .unwrap_or_default(),
-        );
-    };
-
-    let bookings = state
-        .db
-        .list_bookings_by_user(&user_id)
-        .await
-        .unwrap_or_default();
-    let vehicles = state
-        .db
-        .list_vehicles_by_user(&user_id)
-        .await
-        .unwrap_or_default();
-
-    let absences = state
-        .db
-        .list_absences_by_user(&user_id)
-        .await
-        .unwrap_or_default();
-    let credit_transactions = state
-        .db
-        .list_credit_transactions_for_user(auth_user.user_id)
-        .await
-        .unwrap_or_default();
-    let notifications = state
-        .db
-        .list_notifications_by_user(&user_id)
-        .await
-        .unwrap_or_default();
-
-    // Note: password_hash is intentionally excluded from GDPR exports.
-    // Exporting a password hash would allow offline brute-force attacks
-    // against the user's own credential — contrary to the spirit of Art. 15.
-    let export = serde_json::json!({
-        "exported_at": Utc::now().to_rfc3339(),
-        "gdpr_basis": "GDPR Art. 15 — Right of Access",
-        "profile": {
-            "id": user.id,
-            "username": user.username,
-            "email": user.email,
-            "name": user.name,
-            "phone": user.phone,
-            "role": user.role,
-            "created_at": user.created_at,
-            "last_login": user.last_login,
-            "preferences": user.preferences,
-        },
-        "bookings": bookings,
-        "vehicles": vehicles,
-        "absences": absences,
-        "credit_transactions": credit_transactions,
-        "notifications": notifications,
-    });
-
-    let json_str = serde_json::to_string_pretty(&export).unwrap_or_default();
-
-    (
-        StatusCode::OK,
-        [(header::CONTENT_TYPE, "application/json")],
-        json_str,
-    )
-}
-
-/// GDPR Art. 17 — Right to Erasure: anonymize user data, keep booking records for accounting.
-/// Removes PII (name, email, username, password, vehicles) while preserving anonymized booking
-/// records as required by German tax law (§ 147 AO — 10-year retention for accounting records).
-#[utoipa::path(delete, path = "/api/v1/users/me/delete", tag = "Users",
-    summary = "GDPR account deletion (Art. 17)", description = "Anonymizes user PII while preserving booking records.",
-    security(("bearer_auth" = [])),
-    responses((status = 200, description = "Account anonymized"), (status = 404, description = "Not found"))
-)]
-pub async fn gdpr_delete_account(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-) -> (StatusCode, Json<ApiResponse<()>>) {
-    let user_id = auth_user.user_id.to_string();
-    let state_guard = state.read().await;
-
-    // Capture username before anonymization scrubs it
-    let username = state_guard
-        .db
-        .get_user(&user_id)
-        .await
-        .ok()
-        .flatten()
-        .map(|u| u.username)
-        .unwrap_or_default();
-
-    match state_guard.db.anonymize_user(&user_id).await {
-        Ok(true) => {
-            AuditEntry::new(AuditEventType::UserDeleted)
-                .user(auth_user.user_id, &username)
-                .log();
-            (StatusCode::OK, Json(ApiResponse::success(())))
-        }
-        Ok(false) => (
-            StatusCode::NOT_FOUND,
-            Json(ApiResponse::error("NOT_FOUND", "User not found")),
-        ),
-        Err(e) => {
-            tracing::error!("GDPR anonymization failed for {}: {}", user_id, e);
-            (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ApiResponse::error(
-                    "SERVER_ERROR",
-                    "Failed to anonymize account",
-                )),
-            )
-        }
-    }
-}
-
-// ═══════════════════════════════════════════════════════════════════════════════
-// ADMIN — USER MANAGEMENT
-// ═══════════════════════════════════════════════════════════════════════════════
-
-/// Request body for updating a user's role
-#[derive(Debug, Deserialize, utoipa::ToSchema)]
-pub struct UpdateUserRoleRequest {
-    role: String,
-}
-
-/// Request body for updating a user's status
-#[derive(Debug, Deserialize, utoipa::ToSchema)]
-pub struct UpdateUserStatusRequest {
-    status: String,
-}
-
-use admin::AdminUserResponse;
-
-/// `GET /api/v1/admin/users` — list all users (admin only)
-#[utoipa::path(get, path = "/api/v1/admin/users", tag = "Admin",
-    summary = "List all users (admin)", description = "Returns all registered users. Admin only.",
-    security(("bearer_auth" = [])),
-    responses((status = 200, description = "User list"), (status = 403, description = "Forbidden"))
-)]
-#[tracing::instrument(skip(state), fields(admin_id = %auth_user.user_id))]
-pub async fn admin_list_users(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-) -> (StatusCode, Json<ApiResponse<Vec<AdminUserResponse>>>) {
-    let state_guard = state.read().await;
-    if let Err((status, msg)) = check_admin(&state_guard, &auth_user).await {
-        return (status, Json(ApiResponse::error("FORBIDDEN", msg)));
-    }
-
-    match state_guard.db.list_users().await {
-        Ok(users) => {
-            tracing::debug!(count = users.len(), "Admin listed users");
-            let response: Vec<AdminUserResponse> =
-                users.iter().map(AdminUserResponse::from).collect();
-            (StatusCode::OK, Json(ApiResponse::success(response)))
-        }
-        Err(e) => {
-            tracing::error!("Failed to list users: {}", e);
-            (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ApiResponse::error("SERVER_ERROR", "Failed to list users")),
-            )
-        }
-    }
-}
-
-/// `PATCH /api/v1/admin/users/{id}/role` — update a user's role (admin only)
-#[utoipa::path(patch, path = "/api/v1/admin/users/{id}/role", tag = "Admin",
-    summary = "Update user role (admin)", description = "Changes a user's role. Prevents privilege escalation.",
-    security(("bearer_auth" = [])), params(("id" = String, Path, description = "User UUID")),
-    responses((status = 200, description = "Role updated"), (status = 403, description = "Forbidden"), (status = 404, description = "Not found"))
-)]
-#[tracing::instrument(skip(state, req), fields(admin_id = %auth_user.user_id, target_user_id = %id, new_role = %req.role))]
-pub async fn admin_update_user_role(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-    Path(id): Path<String>,
-    Json(req): Json<UpdateUserRoleRequest>,
-) -> (StatusCode, Json<ApiResponse<AdminUserResponse>>) {
-    let state_guard = state.read().await;
-    if let Err((status, msg)) = check_admin(&state_guard, &auth_user).await {
-        return (status, Json(ApiResponse::error("FORBIDDEN", msg)));
-    }
-
-    // Fetch the caller to check their role for privilege escalation prevention
-    let Ok(Some(caller)) = state_guard
-        .db
-        .get_user(&auth_user.user_id.to_string())
-        .await
-    else {
-        return (
-            StatusCode::FORBIDDEN,
-            Json(ApiResponse::error("FORBIDDEN", "Access denied")),
-        );
-    };
-
-    // Only SuperAdmin may promote users to SuperAdmin (prevent privilege escalation)
-    if req.role.as_str() == "superadmin" && caller.role != UserRole::SuperAdmin {
-        return (
-            StatusCode::FORBIDDEN,
-            Json(ApiResponse::error(
-                "FORBIDDEN",
-                "Only a SuperAdmin can assign the SuperAdmin role",
-            )),
-        );
-    }
-
-    let mut user = match state_guard.db.get_user(&id).await {
-        Ok(Some(u)) => u,
-        Ok(None) => {
-            return (
-                StatusCode::NOT_FOUND,
-                Json(ApiResponse::error("NOT_FOUND", "User not found")),
-            );
-        }
-        Err(e) => {
-            tracing::error!("Database error: {}", e);
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ApiResponse::error("SERVER_ERROR", "Internal server error")),
-            );
-        }
-    };
-
-    // Parse role string
-    user.role = match req.role.as_str() {
-        "admin" => UserRole::Admin,
-        "superadmin" => UserRole::SuperAdmin,
-        _ => UserRole::User,
-    };
-    user.updated_at = Utc::now();
-
-    if let Err(e) = state_guard.db.save_user(&user).await {
-        tracing::error!("Failed to update user role: {}", e);
-        return (
-            StatusCode::INTERNAL_SERVER_ERROR,
-            Json(ApiResponse::error("SERVER_ERROR", "Failed to update user")),
-        );
-    }
-
-    let admin_username = state_guard
-        .db
-        .get_user(&auth_user.user_id.to_string())
-        .await
-        .ok()
-        .flatten()
-        .map(|u| u.username)
-        .unwrap_or_default();
-
-    AuditEntry::new(AuditEventType::RoleChanged)
-        .user(auth_user.user_id, &admin_username)
-        .resource("user", &id)
-        .log();
-
-    tracing::info!(
-        admin_id = %auth_user.user_id,
-        target_user_id = %id,
-        new_role = %req.role,
-        "Admin updated user role"
-    );
-
-    (
-        StatusCode::OK,
-        Json(ApiResponse::success(AdminUserResponse::from(&user))),
-    )
-}
-
-/// `PATCH /api/v1/admin/users/{id}/status` — enable or disable a user account (admin only)
-#[utoipa::path(patch, path = "/api/v1/admin/users/{id}/status", tag = "Admin",
-    summary = "Enable or disable a user (admin)", description = "Sets a user's active/inactive status. Admin only.",
-    security(("bearer_auth" = [])), params(("id" = String, Path, description = "User UUID")),
-    responses((status = 200, description = "Updated"), (status = 403, description = "Forbidden"), (status = 404, description = "Not found"))
-)]
-#[tracing::instrument(skip(state, req), fields(admin_id = %auth_user.user_id, target_user_id = %id))]
-pub async fn admin_update_user_status(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-    Path(id): Path<String>,
-    Json(req): Json<UpdateUserStatusRequest>,
-) -> (StatusCode, Json<ApiResponse<AdminUserResponse>>) {
-    let state_guard = state.read().await;
-    if let Err((status, msg)) = check_admin(&state_guard, &auth_user).await {
-        return (status, Json(ApiResponse::error("FORBIDDEN", msg)));
-    }
-
-    let mut user = match state_guard.db.get_user(&id).await {
-        Ok(Some(u)) => u,
-        Ok(None) => {
-            return (
-                StatusCode::NOT_FOUND,
-                Json(ApiResponse::error("NOT_FOUND", "User not found")),
-            );
-        }
-        Err(e) => {
-            tracing::error!("Database error: {}", e);
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ApiResponse::error("SERVER_ERROR", "Internal server error")),
-            );
-        }
-    };
-
-    user.is_active = req.status == "active";
-    user.updated_at = Utc::now();
-
-    if let Err(e) = state_guard.db.save_user(&user).await {
-        tracing::error!("Failed to update user status: {}", e);
-        return (
-            StatusCode::INTERNAL_SERVER_ERROR,
-            Json(ApiResponse::error("SERVER_ERROR", "Failed to update user")),
-        );
-    }
-
-    // Revoke all sessions when a user is disabled
-    if !user.is_active {
-        if let Err(e) = state_guard.db.delete_sessions_by_user(user.id).await {
-            tracing::error!("Failed to revoke sessions for disabled user {}: {}", id, e);
-        }
-    }
-
-    let event_type = if user.is_active {
-        AuditEventType::UserActivated
-    } else {
-        AuditEventType::UserDeactivated
-    };
-    let audit = AuditEntry::new(event_type)
-        .user(auth_user.user_id, "admin")
-        .resource("user", &id)
-        .details(serde_json::json!({ "new_status": req.status }))
-        .log();
-    audit.persist(&state_guard.db).await;
-
-    tracing::info!(
-        admin_id = %auth_user.user_id,
-        target_user_id = %id,
-        new_status = %req.status,
-        "Admin updated user status"
-    );
-
-    (
-        StatusCode::OK,
-        Json(ApiResponse::success(AdminUserResponse::from(&user))),
-    )
-}
-
-/// `DELETE /api/v1/admin/users/{id}` — delete a user account (admin only, GDPR anonymize)
-#[utoipa::path(delete, path = "/api/v1/admin/users/{id}", tag = "Admin",
-    summary = "Delete user (admin)", description = "Anonymizes user data per GDPR. Admin only.",
-    security(("bearer_auth" = [])), params(("id" = String, Path, description = "User UUID")),
-    responses((status = 200, description = "Deleted"), (status = 403, description = "Forbidden"), (status = 404, description = "Not found"))
-)]
-#[tracing::instrument(skip(state), fields(admin_id = %auth_user.user_id, target_user_id = %id))]
-pub async fn admin_delete_user(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-    Path(id): Path<String>,
-) -> (StatusCode, Json<ApiResponse<()>>) {
-    let state_guard = state.read().await;
-    if let Err((status, msg)) = check_admin(&state_guard, &auth_user).await {
-        return (status, Json(ApiResponse::error("FORBIDDEN", msg)));
-    }
-
-    // Prevent admin from deleting their own account via admin panel
-    if id == auth_user.user_id.to_string() {
-        return (
-            StatusCode::BAD_REQUEST,
-            Json(ApiResponse::error(
-                "CANNOT_DELETE_SELF",
-                "You cannot delete your own account",
-            )),
-        );
-    }
-
-    let admin_username = state_guard
-        .db
-        .get_user(&auth_user.user_id.to_string())
-        .await
-        .ok()
-        .flatten()
-        .map(|u| u.username)
-        .unwrap_or_default();
-
-    match state_guard.db.anonymize_user(&id).await {
-        Ok(true) => {
-            AuditEntry::new(AuditEventType::UserDeleted)
-                .user(auth_user.user_id, &admin_username)
-                .resource("user", &id)
-                .log();
-            tracing::info!(
-                admin_id = %auth_user.user_id,
-                target_user_id = %id,
-                "Admin anonymized user"
-            );
-            (StatusCode::OK, Json(ApiResponse::success(())))
-        }
-        Ok(false) => (
-            StatusCode::NOT_FOUND,
-            Json(ApiResponse::error("NOT_FOUND", "User not found")),
-        ),
-        Err(e) => {
-            tracing::error!("Failed to anonymize user {}: {}", id, e);
-            (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ApiResponse::error("SERVER_ERROR", "Failed to delete user")),
-            )
-        }
-    }
-}
-
-// ═══════════════════════════════════════════════════════════════════════════════
-// ADMIN — BOOKING MANAGEMENT
-// ═══════════════════════════════════════════════════════════════════════════════
-
-/// Response type for admin booking listing (includes user details)
-#[derive(Debug, Serialize)]
-pub struct AdminBookingResponse {
-    id: String,
-    user_id: String,
-    user_name: String,
-    user_email: String,
-    lot_id: String,
-    lot_name: String,
-    slot_id: String,
-    slot_number: String,
-    vehicle_plate: String,
-    start_time: chrono::DateTime<Utc>,
-    end_time: chrono::DateTime<Utc>,
-    status: String,
-    created_at: chrono::DateTime<Utc>,
-}
-
-/// `GET /api/v1/admin/bookings` — list all bookings (admin only)
-#[utoipa::path(get, path = "/api/v1/admin/bookings", tag = "Admin",
-    summary = "List all bookings (admin)", description = "Returns all bookings with enriched details. Admin only.",
-    security(("bearer_auth" = [])),
-    responses((status = 200, description = "All bookings"), (status = 403, description = "Forbidden"))
-)]
-pub async fn admin_list_bookings(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-) -> (StatusCode, Json<ApiResponse<Vec<AdminBookingResponse>>>) {
-    let state_guard = state.read().await;
-    if let Err((status, msg)) = check_admin(&state_guard, &auth_user).await {
-        return (status, Json(ApiResponse::error("FORBIDDEN", msg)));
-    }
-
-    let bookings = match state_guard.db.list_bookings().await {
-        Ok(b) => b,
-        Err(e) => {
-            tracing::error!("Failed to list bookings: {}", e);
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ApiResponse::error(
-                    "SERVER_ERROR",
-                    "Failed to list bookings",
-                )),
-            );
-        }
-    };
-
-    // Batch-load all users and lots upfront to avoid N+1 queries
-    let all_users = state_guard.db.list_users().await.unwrap_or_default();
-    let user_map: std::collections::HashMap<String, _> = all_users
-        .into_iter()
-        .map(|u| (u.id.to_string(), u))
-        .collect();
-
-    let all_lots = state_guard.db.list_parking_lots().await.unwrap_or_default();
-    let lot_map: std::collections::HashMap<String, _> = all_lots
-        .into_iter()
-        .map(|l| (l.id.to_string(), l))
-        .collect();
-
-    let mut response = Vec::with_capacity(bookings.len());
-    for booking in bookings {
-        let (user_name, user_email) = match user_map.get(&booking.user_id.to_string()) {
-            Some(u) => (u.name.clone(), u.email.clone()),
-            None => (booking.user_id.to_string(), String::new()),
-        };
-
-        let lot_name = match lot_map.get(&booking.lot_id.to_string()) {
-            Some(l) => l.name.clone(),
-            None => booking.lot_id.to_string(),
-        };
-
-        response.push(AdminBookingResponse {
-            id: booking.id.to_string(),
-            user_id: booking.user_id.to_string(),
-            user_name,
-            user_email,
-            lot_id: booking.lot_id.to_string(),
-            lot_name,
-            slot_id: booking.slot_id.to_string(),
-            slot_number: booking.slot_number.to_string(),
-            vehicle_plate: booking.vehicle.license_plate.clone(),
-            start_time: booking.start_time,
-            end_time: booking.end_time,
-            status: format!("{:?}", booking.status).to_lowercase(),
-            created_at: booking.created_at,
-        });
-    }
-
-    (StatusCode::OK, Json(ApiResponse::success(response)))
-}
-
-// ═══════════════════════════════════════════════════════════════════════════════
-// ADMIN REPORTS
-// ═══════════════════════════════════════════════════════════════════════════════
-
-/// Dashboard stats response
-#[derive(Debug, Serialize)]
-pub struct AdminStatsResponse {
-    total_users: u64,
-    total_lots: u64,
-    total_slots: u64,
-    total_bookings: u64,
-    active_bookings: u64,
-    occupancy_percent: f64,
-}
-
-/// `GET /api/v1/admin/stats` — dashboard stats
-#[utoipa::path(get, path = "/api/v1/admin/stats", tag = "Admin",
-    summary = "Admin dashboard statistics",
-    description = "Returns aggregated system stats.",
-    security(("bearer_auth" = [])),
-    responses((status = 200, description = "Success"))
-)]
-#[tracing::instrument(skip(state), fields(admin_id = %auth_user.user_id))]
-pub async fn admin_stats(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-) -> (StatusCode, Json<ApiResponse<AdminStatsResponse>>) {
-    let state_guard = state.read().await;
-    if let Err((status, msg)) = check_admin(&state_guard, &auth_user).await {
-        return (status, Json(ApiResponse::error("FORBIDDEN", msg)));
-    }
-
-    let db_stats = state_guard
-        .db
-        .stats()
-        .await
-        .unwrap_or(crate::db::DatabaseStats {
-            users: 0,
-            bookings: 0,
-            parking_lots: 0,
-            slots: 0,
-            sessions: 0,
-            vehicles: 0,
-        });
-
-    // Count active bookings
-    let active_bookings = state_guard
-        .db
-        .list_bookings()
-        .await
-        .map(|bookings| {
-            bookings
-                .iter()
-                .filter(|b| {
-                    b.status == BookingStatus::Confirmed || b.status == BookingStatus::Active
-                })
-                .count() as u64
-        })
-        .unwrap_or(0);
-
-    #[allow(clippy::cast_precision_loss)]
-    let occupancy = if db_stats.slots > 0 {
-        (active_bookings as f64 / db_stats.slots as f64) * 100.0
-    } else {
-        0.0
-    };
-
-    (
-        StatusCode::OK,
-        Json(ApiResponse::success(AdminStatsResponse {
-            total_users: db_stats.users,
-            total_lots: db_stats.parking_lots,
-            total_slots: db_stats.slots,
-            total_bookings: db_stats.bookings,
-            active_bookings,
-            occupancy_percent: (occupancy * 100.0).round() / 100.0,
-        })),
-    )
-}
-
-/// Query params for reports
-#[derive(Debug, Deserialize)]
-pub struct ReportsQuery {
-    days: Option<i64>,
-}
-
-/// Booking stats by day
-#[derive(Debug, Serialize)]
-pub struct DailyBookingStat {
-    date: String,
-    count: usize,
-}
-
-/// `GET /api/v1/admin/reports` — booking stats by day for last N days
-#[utoipa::path(get, path = "/api/v1/admin/reports", tag = "Admin",
-    summary = "Booking reports (admin)",
-    description = "Returns daily booking stats.",
-    security(("bearer_auth" = [])),
-    responses((status = 200, description = "Success"))
-)]
-pub async fn admin_reports(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-    Query(query): Query<ReportsQuery>,
-) -> (StatusCode, Json<ApiResponse<Vec<DailyBookingStat>>>) {
-    let state_guard = state.read().await;
-    if let Err((status, msg)) = check_admin(&state_guard, &auth_user).await {
-        return (status, Json(ApiResponse::error("FORBIDDEN", msg)));
-    }
-
-    let days = query.days.unwrap_or(30);
-    let cutoff = Utc::now() - TimeDelta::days(days);
-
-    let bookings = state_guard.db.list_bookings().await.unwrap_or_default();
-
-    // Group by date
-    let mut by_date: std::collections::BTreeMap<String, usize> = std::collections::BTreeMap::new();
-    for b in &bookings {
-        if b.created_at >= cutoff {
-            let date = b.created_at.format("%Y-%m-%d").to_string();
-            *by_date.entry(date).or_insert(0) += 1;
-        }
-    }
-
-    let daily_stats: Vec<DailyBookingStat> = by_date
-        .into_iter()
-        .map(|(date, count)| DailyBookingStat { date, count })
-        .collect();
-
-    (StatusCode::OK, Json(ApiResponse::success(daily_stats)))
-}
-
-/// Heatmap cell: booking count by weekday x hour
-#[derive(Debug, Serialize)]
-pub struct HeatmapCell {
-    weekday: u32,
-    hour: u32,
-    count: usize,
-}
-
-/// `GET /api/v1/admin/heatmap` — booking counts by weekday x hour
-#[utoipa::path(get, path = "/api/v1/admin/heatmap", tag = "Admin",
-    summary = "Booking heatmap (admin)",
-    description = "Returns booking counts by weekday and hour.",
-    security(("bearer_auth" = [])),
-    responses((status = 200, description = "Success"))
-)]
-pub async fn admin_heatmap(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-) -> (StatusCode, Json<ApiResponse<Vec<HeatmapCell>>>) {
-    let state_guard = state.read().await;
-    if let Err((status, msg)) = check_admin(&state_guard, &auth_user).await {
-        return (status, Json(ApiResponse::error("FORBIDDEN", msg)));
-    }
-
-    let bookings = state_guard.db.list_bookings().await.unwrap_or_default();
-
-    // Build 7x24 grid (weekday 0=Mon .. 6=Sun, hour 0..23)
-    let mut grid = [[0usize; 24]; 7];
-    for b in &bookings {
-        let weekday = b.start_time.weekday().num_days_from_monday() as usize;
-        let hour = b.start_time.hour() as usize;
-        if weekday < 7 && hour < 24 {
-            grid[weekday][hour] += 1;
-        }
-    }
-
-    let cells: Vec<HeatmapCell> = grid
-        .iter()
-        .enumerate()
-        .flat_map(|(wd, hours)| {
-            hours
-                .iter()
-                .enumerate()
-                .map(move |(h, &count)| HeatmapCell {
-                    weekday: u32::try_from(wd).unwrap_or(0),
-                    hour: u32::try_from(h).unwrap_or(0),
-                    count,
-                })
-        })
-        .collect();
-
-    (StatusCode::OK, Json(ApiResponse::success(cells)))
-}
-
-// ═══════════════════════════════════════════════════════════════════════════════
-// AUDIT LOG
-// ═══════════════════════════════════════════════════════════════════════════════
-
-/// `GET /api/v1/admin/audit-log` — list recent audit entries
-#[utoipa::path(get, path = "/api/v1/admin/audit-log", tag = "Admin",
-    summary = "Audit log (admin)",
-    description = "Returns recent audit log entries.",
-    security(("bearer_auth" = [])),
-    responses((status = 200, description = "Success"))
-)]
-pub async fn admin_audit_log(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-    axum::extract::Query(params): axum::extract::Query<std::collections::HashMap<String, String>>,
-) -> (StatusCode, Json<ApiResponse<Vec<crate::db::AuditLogEntry>>>) {
-    let state_guard = state.read().await;
-    if let Err((status, msg)) = check_admin(&state_guard, &auth_user).await {
-        return (status, Json(ApiResponse::error("FORBIDDEN", msg)));
-    }
-
-    let limit = params
-        .get("limit")
-        .and_then(|v| v.parse().ok())
-        .unwrap_or(100usize)
-        .min(500);
-
-    match state_guard.db.list_audit_log(limit).await {
-        Ok(entries) => (StatusCode::OK, Json(ApiResponse::success(entries))),
-        Err(e) => {
-            tracing::error!("Failed to list audit log: {e}");
-            (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ApiResponse::error(
-                    "SERVER_ERROR",
-                    "Failed to list audit log",
-                )),
-            )
-        }
-    }
-}
-
-// ═══════════════════════════════════════════════════════════════════════════════
-// CHANGE PASSWORD
-// ═══════════════════════════════════════════════════════════════════════════════
-
-/// Request body for password change
-#[derive(Debug, Deserialize, utoipa::ToSchema)]
-pub struct ChangePasswordRequest {
-    current_password: String,
-    new_password: String,
-}
-
-/// `PATCH /api/v1/users/me/password` — authenticated user changes their own password
-#[utoipa::path(patch, path = "/api/v1/users/me/password", tag = "Users",
-    summary = "Change password",
-    description = "Changes the authenticated user password.",
-    security(("bearer_auth" = [])),
-    responses((status = 200, description = "Success"))
-)]
-#[tracing::instrument(skip(state, req), fields(user_id = %auth_user.user_id))]
-pub async fn change_password(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-    Json(req): Json<ChangePasswordRequest>,
-) -> (StatusCode, Json<ApiResponse<()>>) {
-    // Validate new password length
-    if req.new_password.len() < 8 {
-        return (
-            StatusCode::BAD_REQUEST,
-            Json(ApiResponse::error(
-                "VALIDATION_ERROR",
-                "New password must be at least 8 characters",
-            )),
-        );
-    }
-
-    let state_guard = state.read().await;
-    let user = match state_guard
-        .db
-        .get_user(&auth_user.user_id.to_string())
-        .await
-    {
-        Ok(Some(u)) => u,
-        Ok(None) => {
-            return (
-                StatusCode::NOT_FOUND,
-                Json(ApiResponse::error("NOT_FOUND", "User not found")),
-            );
-        }
-        Err(e) => {
-            tracing::error!("Database error: {}", e);
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ApiResponse::error("SERVER_ERROR", "Internal server error")),
-            );
-        }
-    };
-
-    // Verify current password
-    if !verify_password(&req.current_password, &user.password_hash).await {
-        return (
-            StatusCode::UNAUTHORIZED,
-            Json(ApiResponse::error(
-                "INVALID_PASSWORD",
-                "Current password is incorrect",
-            )),
-        );
-    }
-
-    // Hash new password
-    let new_hash = match hash_password_simple(&req.new_password).await {
-        Ok(h) => h,
-        Err(e) => {
-            tracing::error!("Password hashing failed: {}", e);
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ApiResponse::error("SERVER_ERROR", "Internal server error")),
-            );
-        }
-    };
-
-    // Update user
-    let mut updated_user = user;
-    updated_user.password_hash = new_hash;
-    updated_user.updated_at = Utc::now();
-
-    if let Err(e) = state_guard.db.save_user(&updated_user).await {
-        tracing::error!("Failed to save user: {}", e);
-        return (
-            StatusCode::INTERNAL_SERVER_ERROR,
-            Json(ApiResponse::error(
-                "SERVER_ERROR",
-                "Failed to update password",
-            )),
-        );
-    }
-
-    // Invalidate all existing sessions for this user — a password change must
-    // force re-authentication on every device (issue #116).
-    if let Err(e) = state_guard
-        .db
-        .delete_sessions_by_user(auth_user.user_id)
-        .await
-    {
-        tracing::warn!(
-            user_id = %auth_user.user_id,
-            error = %e,
-            "Failed to invalidate sessions after password change"
-        );
-    }
-
-    crate::audit::AuditEntry::new(crate::audit::AuditEventType::PasswordChanged)
-        .user(auth_user.user_id, &updated_user.username)
-        .log();
-
-    (StatusCode::OK, Json(ApiResponse::success(())))
-}
-
-// ═══════════════════════════════════════════════════════════════════════════════
-// PUBLIC OCCUPANCY
-// ═══════════════════════════════════════════════════════════════════════════════
-
-/// Occupancy info for a single lot
-#[derive(Debug, Serialize)]
-pub struct LotOccupancy {
-    lot_id: String,
-    lot_name: String,
-    total_slots: i32,
-    occupied_slots: i32,
-    available_slots: i32,
-}
-
-/// `GET /api/v1/public/occupancy` — public JSON occupancy data
-#[utoipa::path(get, path = "/api/v1/public/occupancy", tag = "Public",
-    summary = "Public lot occupancy",
-    description = "Returns real-time occupancy. No auth required.",
-    responses((status = 200, description = "Success"))
-)]
-pub async fn public_occupancy(
-    State(state): State<SharedState>,
-) -> (StatusCode, Json<ApiResponse<Vec<LotOccupancy>>>) {
-    let state_guard = state.read().await;
-
-    let lots = match state_guard.db.list_parking_lots().await {
-        Ok(l) => l,
-        Err(e) => {
-            tracing::error!("Failed to list lots for occupancy: {}", e);
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ApiResponse::error(
-                    "SERVER_ERROR",
-                    "Failed to get occupancy",
-                )),
-            );
-        }
-    };
-
-    // Count active bookings per lot
-    let now = Utc::now();
-    let bookings = state_guard.db.list_bookings().await.unwrap_or_default();
-
-    let mut occupancy = Vec::with_capacity(lots.len());
-    for lot in &lots {
-        let occupied = i32::try_from(
-            bookings
-                .iter()
-                .filter(|b| {
-                    b.lot_id == lot.id
-                        && b.start_time <= now
-                        && b.end_time >= now
-                        && matches!(
-                            b.status,
-                            parkhub_common::BookingStatus::Confirmed
-                                | parkhub_common::BookingStatus::Active
-                        )
-                })
-                .count(),
-        )
-        .unwrap_or(i32::MAX);
-
-        let available = (lot.total_slots - occupied).max(0);
-
-        occupancy.push(LotOccupancy {
-            lot_id: lot.id.to_string(),
-            lot_name: lot.name.clone(),
-            total_slots: lot.total_slots,
-            occupied_slots: occupied,
-            available_slots: available,
-        });
-    }
-
-    (StatusCode::OK, Json(ApiResponse::success(occupancy)))
-}
-
-/// `GET /api/v1/public/display` — simplified HTML for parking displays
-#[utoipa::path(get, path = "/api/v1/public/display", tag = "Public",
-    summary = "Public display HTML",
-    description = "Returns minimal HTML for digital signage.",
-    responses((status = 200, description = "Success"))
-)]
-pub async fn public_display(State(state): State<SharedState>) -> impl axum::response::IntoResponse {
-    let state_guard = state.read().await;
-
-    let lots = state_guard.db.list_parking_lots().await.unwrap_or_default();
-    let now = Utc::now();
-    let bookings = state_guard.db.list_bookings().await.unwrap_or_default();
-
-    let mut html = String::from(
-        r#"<!DOCTYPE html>
-<html lang="en">
-<head>
-<meta charset="UTF-8">
-<meta name="viewport" content="width=device-width, initial-scale=1.0">
-<meta http-equiv="refresh" content="30">
-<title>ParkHub — Parking Availability</title>
-<style>
-  body { font-family: system-ui, sans-serif; margin: 0; padding: 2rem; background: #1a1a2e; color: #eee; }
-  h1 { text-align: center; margin-bottom: 2rem; }
-  .lots { display: flex; flex-wrap: wrap; gap: 1.5rem; justify-content: center; }
-  .lot { background: #16213e; border-radius: 12px; padding: 1.5rem 2rem; min-width: 220px; text-align: center; }
-  .lot-name { font-size: 1.2rem; font-weight: 600; margin-bottom: 0.5rem; }
-  .available { font-size: 3rem; font-weight: 700; }
-  .available.green { color: #4ade80; }
-  .available.yellow { color: #facc15; }
-  .available.red { color: #f87171; }
-  .label { font-size: 0.9rem; color: #94a3b8; }
-</style>
-</head>
-<body>
-<h1>Parking Availability</h1>
-<div class="lots">
-"#,
-    );
-
-    for lot in &lots {
-        let occupied = i32::try_from(
-            bookings
-                .iter()
-                .filter(|b| {
-                    b.lot_id == lot.id
-                        && b.start_time <= now
-                        && b.end_time >= now
-                        && matches!(
-                            b.status,
-                            parkhub_common::BookingStatus::Confirmed
-                                | parkhub_common::BookingStatus::Active
-                        )
-                })
-                .count(),
-        )
-        .unwrap_or(i32::MAX);
-
-        let available = (lot.total_slots - occupied).max(0);
-        let pct = if lot.total_slots > 0 {
-            (f64::from(available) / f64::from(lot.total_slots)) * 100.0
-        } else {
-            0.0
-        };
-        let color_class = if pct > 30.0 {
-            "green"
-        } else if pct > 10.0 {
-            "yellow"
-        } else {
-            "red"
-        };
-
-        {
-            let _ = write!(
-                html,
-                r#"<div class="lot">
-  <div class="lot-name">{}</div>
-  <div class="available {}">{}</div>
-  <div class="label">of {} available</div>
-</div>
-"#,
-                crate::utils::html_escape(&lot.name),
-                color_class,
-                available,
-                lot.total_slots
-            );
-        }
-    }
-
-    html.push_str("</div>\n</body>\n</html>\n");
-
-    (
-        StatusCode::OK,
-        [(header::CONTENT_TYPE, "text/html; charset=utf-8")],
-        html,
-    )
-}
-
-// ═══════════════════════════════════════════════════════════════════════════════
-// USER STATS
-// ═══════════════════════════════════════════════════════════════════════════════
-
-/// `GET /api/v1/user/stats` — authenticated user's personal statistics
-#[utoipa::path(get, path = "/api/v1/user/stats", tag = "Users",
-    summary = "User personal statistics",
-    description = "Returns personal parking stats.",
-    security(("bearer_auth" = [])),
-    responses((status = 200, description = "Success"))
-)]
-pub async fn user_stats(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-) -> (StatusCode, Json<ApiResponse<serde_json::Value>>) {
-    let state_guard = state.read().await;
-    let uid = auth_user.user_id.to_string();
-
-    let Ok(Some(user)) = state_guard.db.get_user(&uid).await else {
-        return (
-            StatusCode::NOT_FOUND,
-            Json(ApiResponse::error("NOT_FOUND", "User not found")),
-        );
-    };
-
-    let bookings = state_guard
-        .db
-        .list_bookings_by_user(&uid)
-        .await
-        .unwrap_or_default();
-
-    let total_bookings = bookings.len();
-    let active_bookings = bookings
-        .iter()
-        .filter(|b| {
-            matches!(
-                b.status,
-                BookingStatus::Confirmed | BookingStatus::Active | BookingStatus::Pending
-            )
-        })
-        .count();
-    let cancelled_bookings = bookings
-        .iter()
-        .filter(|b| b.status == BookingStatus::Cancelled)
-        .count();
-
-    // Sum credits spent from deduction transactions
-    let total_credits_spent = state_guard
-        .db
-        .list_credit_transactions_for_user(auth_user.user_id)
-        .await
-        .unwrap_or_default()
-        .iter()
-        .filter(|tx| tx.transaction_type == CreditTransactionType::Deduction)
-        .map(|tx| i64::from(tx.amount.abs()))
-        .sum::<i64>();
-
-    // Find favorite lot by most bookings
-    let favorite_lot = {
-        let mut lot_counts: std::collections::HashMap<Uuid, usize> =
-            std::collections::HashMap::new();
-        for b in &bookings {
-            *lot_counts.entry(b.lot_id).or_insert(0) += 1;
-        }
-        if let Some((&lot_id, _)) = lot_counts.iter().max_by_key(|(_, &c)| c) {
-            state_guard
-                .db
-                .get_parking_lot(&lot_id.to_string())
-                .await
-                .ok()
-                .flatten()
-                .map_or_else(|| "Unknown".to_string(), |l| l.name)
-        } else {
-            "None".to_string()
-        }
-    };
-
-    (
-        StatusCode::OK,
-        Json(ApiResponse::success(serde_json::json!({
-            "total_bookings": total_bookings,
-            "active_bookings": active_bookings,
-            "cancelled_bookings": cancelled_bookings,
-            "total_credits_spent": total_credits_spent,
-            "favorite_lot": favorite_lot,
-            "member_since": user.created_at,
-        }))),
-    )
-}
-
-// ═══════════════════════════════════════════════════════════════════════════════
-// USER PREFERENCES
-// ═══════════════════════════════════════════════════════════════════════════════
-
-/// `GET /api/v1/user/preferences` — return current user's preferences
-#[utoipa::path(get, path = "/api/v1/user/preferences", tag = "Users",
-    summary = "Get user preferences",
-    description = "Returns user preferences.",
-    security(("bearer_auth" = [])),
-    responses((status = 200, description = "Success"))
-)]
-pub async fn get_user_preferences(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-) -> (StatusCode, Json<ApiResponse<serde_json::Value>>) {
-    let state_guard = state.read().await;
-
-    match state_guard
-        .db
-        .get_user(&auth_user.user_id.to_string())
-        .await
-    {
-        Ok(Some(user)) => (
-            StatusCode::OK,
-            Json(ApiResponse::success(serde_json::json!({
-                "language": user.preferences.language,
-                "theme": user.preferences.theme,
-                "notifications_enabled": user.preferences.notifications_enabled,
-                "email_reminders": user.preferences.email_reminders,
-                "default_duration_minutes": user.preferences.default_duration_minutes,
-            }))),
-        ),
-        _ => (
-            StatusCode::NOT_FOUND,
-            Json(ApiResponse::error("NOT_FOUND", "User not found")),
-        ),
-    }
-}
-
-/// Request body for updating user preferences
-#[derive(Debug, Deserialize, utoipa::ToSchema)]
-pub struct UpdatePreferencesRequest {
-    language: Option<String>,
-    theme: Option<String>,
-    notifications_enabled: Option<bool>,
-    email_reminders: Option<bool>,
-    default_duration_minutes: Option<i32>,
-}
-
-/// `PUT /api/v1/user/preferences` — update preferences
-#[utoipa::path(put, path = "/api/v1/user/preferences", tag = "Users",
-    summary = "Update user preferences",
-    description = "Updates user preferences.",
-    security(("bearer_auth" = [])),
-    responses((status = 200, description = "Success"))
-)]
-pub async fn update_user_preferences(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-    Json(req): Json<UpdatePreferencesRequest>,
-) -> (StatusCode, Json<ApiResponse<serde_json::Value>>) {
-    let state_guard = state.read().await;
-
-    let Ok(Some(mut user)) = state_guard
-        .db
-        .get_user(&auth_user.user_id.to_string())
-        .await
-    else {
-        return (
-            StatusCode::NOT_FOUND,
-            Json(ApiResponse::error("NOT_FOUND", "User not found")),
-        );
-    };
-
-    if let Some(lang) = req.language {
-        user.preferences.language = lang;
-    }
-    if let Some(theme) = req.theme {
-        user.preferences.theme = theme;
-    }
-    if let Some(notif) = req.notifications_enabled {
-        user.preferences.notifications_enabled = notif;
-    }
-    if let Some(email) = req.email_reminders {
-        user.preferences.email_reminders = email;
-    }
-    if let Some(dur) = req.default_duration_minutes {
-        user.preferences.default_duration_minutes = Some(dur);
-    }
-    user.updated_at = Utc::now();
-
-    if let Err(e) = state_guard.db.save_user(&user).await {
-        tracing::error!("Failed to save preferences: {}", e);
-        return (
-            StatusCode::INTERNAL_SERVER_ERROR,
-            Json(ApiResponse::error(
-                "SERVER_ERROR",
-                "Failed to save preferences",
-            )),
-        );
-    }
-
-    (
-        StatusCode::OK,
-        Json(ApiResponse::success(serde_json::json!({
-            "language": user.preferences.language,
-            "theme": user.preferences.theme,
-            "notifications_enabled": user.preferences.notifications_enabled,
-            "email_reminders": user.preferences.email_reminders,
-            "default_duration_minutes": user.preferences.default_duration_minutes,
-        }))),
-    )
-}
-
-// ═══════════════════════════════════════════════════════════════════════════════
-// ADMIN: DATABASE RESET
-// ═══════════════════════════════════════════════════════════════════════════════
-
-/// Request body for database reset confirmation
-#[derive(Debug, Deserialize, utoipa::ToSchema)]
-pub struct AdminResetRequest {
-    confirm: String,
-}
-
-/// `POST /api/v1/admin/reset` — wipe all data (admin only)
-#[utoipa::path(post, path = "/api/v1/admin/reset", tag = "Admin",
-    summary = "Reset database (admin)",
-    description = "Wipes all data. Destructive. Admin only.",
-    security(("bearer_auth" = [])),
-    responses((status = 200, description = "Success"))
-)]
-pub async fn admin_reset(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-    Json(req): Json<AdminResetRequest>,
-) -> (StatusCode, Json<ApiResponse<()>>) {
-    let state_guard = state.write().await;
-    if let Err((status, msg)) = check_admin(&state_guard, &auth_user).await {
-        return (status, Json(ApiResponse::error("FORBIDDEN", msg)));
-    }
-
-    if req.confirm != "RESET" {
-        return (
-            StatusCode::BAD_REQUEST,
-            Json(ApiResponse::error(
-                "CONFIRMATION_REQUIRED",
-                "Body must contain {\"confirm\": \"RESET\"}",
-            )),
-        );
-    }
-
-    // Capture admin info before wipe
-    let Ok(Some(admin)) = state_guard
-        .db
-        .get_user(&auth_user.user_id.to_string())
-        .await
-    else {
-        return (
-            StatusCode::INTERNAL_SERVER_ERROR,
-            Json(ApiResponse::error(
-                "SERVER_ERROR",
-                "Failed to read admin user before reset",
-            )),
-        );
-    };
-
-    if let Err(e) = state_guard.db.clear_all_data().await {
-        tracing::error!("Database reset failed: {}", e);
-        return (
-            StatusCode::INTERNAL_SERVER_ERROR,
-            Json(ApiResponse::error(
-                "SERVER_ERROR",
-                "Failed to reset database",
-            )),
-        );
-    }
-
-    // Re-create the admin user who triggered the reset
-    let admin_user = User {
-        id: admin.id,
-        username: admin.username.clone(),
-        email: admin.email.clone(),
-        name: admin.name.clone(),
-        password_hash: admin.password_hash,
-        role: admin.role,
-        is_active: true,
-        phone: admin.phone,
-        picture: admin.picture,
-        preferences: admin.preferences,
-        credits_balance: 0,
-        credits_monthly_quota: 0,
-        credits_last_refilled: None,
-        created_at: Utc::now(),
-        updated_at: Utc::now(),
-        last_login: None,
-    };
-
-    if let Err(e) = state_guard.db.save_user(&admin_user).await {
-        tracing::error!("Failed to re-create admin after reset: {}", e);
-        return (
-            StatusCode::INTERNAL_SERVER_ERROR,
-            Json(ApiResponse::error(
-                "SERVER_ERROR",
-                "Database reset succeeded but admin re-creation failed",
-            )),
-        );
-    }
-
-    AuditEntry::new(AuditEventType::ConfigChanged)
-        .user(auth_user.user_id, &admin_user.username)
-        .details(serde_json::json!({"action": "database_reset"}))
-        .log();
-
-    tracing::warn!(
-        admin = %admin_user.username,
-        "Database reset completed"
-    );
-
-    (StatusCode::OK, Json(ApiResponse::success(())))
-}
-
-// ═══════════════════════════════════════════════════════════════════════════════
-// ADMIN: AUTO-RELEASE SETTINGS
-// ═══════════════════════════════════════════════════════════════════════════════
-
-/// `GET /api/v1/admin/settings/auto-release` — return auto-release config
-#[utoipa::path(
-    get,
-    path = "/api/v1/admin/settings/auto-release",
-    tag = "Admin",
-    summary = "Get auto-release settings",
-    description = "Return the auto-release timing configuration. Admin only.",
-    security(("bearer_auth" = []))
-)]
-pub async fn admin_get_auto_release(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-) -> (StatusCode, Json<ApiResponse<serde_json::Value>>) {
-    let state_guard = state.read().await;
-    if let Err((status, msg)) = check_admin(&state_guard, &auth_user).await {
-        return (status, Json(ApiResponse::error("FORBIDDEN", msg)));
-    }
-
-    let enabled = read_admin_setting(&state_guard.db, "auto_release_enabled").await;
-    let minutes = read_admin_setting(&state_guard.db, "auto_release_minutes").await;
-
-    (
-        StatusCode::OK,
-        Json(ApiResponse::success(serde_json::json!({
-            "auto_release_enabled": enabled.parse::<bool>().unwrap_or(false),
-            "auto_release_minutes": minutes.parse::<i32>().unwrap_or(30),
-        }))),
-    )
-}
-
-/// Request body for auto-release settings update
-#[derive(Debug, Deserialize, utoipa::ToSchema)]
-pub struct AutoReleaseSettingsRequest {
-    auto_release_enabled: Option<bool>,
-    auto_release_minutes: Option<i32>,
-}
-
-/// `PUT /api/v1/admin/settings/auto-release` — update auto-release timing
-#[utoipa::path(
-    put,
-    path = "/api/v1/admin/settings/auto-release",
-    tag = "Admin",
-    summary = "Update auto-release settings",
-    description = "Update auto-release timing for unclaimed bookings. Admin only.",
-    security(("bearer_auth" = []))
-)]
-pub async fn admin_update_auto_release(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-    Json(req): Json<AutoReleaseSettingsRequest>,
-) -> (StatusCode, Json<ApiResponse<serde_json::Value>>) {
-    let state_guard = state.read().await;
-    if let Err((status, msg)) = check_admin(&state_guard, &auth_user).await {
-        return (status, Json(ApiResponse::error("FORBIDDEN", msg)));
-    }
-
-    if let Some(enabled) = req.auto_release_enabled {
-        if let Err(e) = state_guard
-            .db
-            .set_setting("auto_release_enabled", &enabled.to_string())
-            .await
-        {
-            tracing::error!("Failed to save auto_release_enabled: {}", e);
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ApiResponse::error("SERVER_ERROR", "Failed to save setting")),
-            );
-        }
-    }
-
-    if let Some(minutes) = req.auto_release_minutes {
-        if minutes < 1 {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(ApiResponse::error(
-                    "INVALID_INPUT",
-                    "auto_release_minutes must be >= 1",
-                )),
-            );
-        }
-        if let Err(e) = state_guard
-            .db
-            .set_setting("auto_release_minutes", &minutes.to_string())
-            .await
-        {
-            tracing::error!("Failed to save auto_release_minutes: {}", e);
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ApiResponse::error("SERVER_ERROR", "Failed to save setting")),
-            );
-        }
-    }
-
-    // Return updated values
-    let enabled = read_admin_setting(&state_guard.db, "auto_release_enabled").await;
-    let minutes = read_admin_setting(&state_guard.db, "auto_release_minutes").await;
-
-    (
-        StatusCode::OK,
-        Json(ApiResponse::success(serde_json::json!({
-            "auto_release_enabled": enabled.parse::<bool>().unwrap_or(false),
-            "auto_release_minutes": minutes.parse::<i32>().unwrap_or(30),
-        }))),
-    )
-}
-
-// ═══════════════════════════════════════════════════════════════════════════════
-// ADMIN: EMAIL SETTINGS
-// ═══════════════════════════════════════════════════════════════════════════════
-
-/// `GET /api/v1/admin/settings/email` — return SMTP config (password masked)
-#[utoipa::path(
-    get,
-    path = "/api/v1/admin/settings/email",
-    tag = "Admin",
-    summary = "Get email settings",
-    description = "Return SMTP configuration (password masked). Admin only.",
-    security(("bearer_auth" = []))
-)]
-pub async fn admin_get_email_settings(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-) -> (StatusCode, Json<ApiResponse<serde_json::Value>>) {
-    let state_guard = state.read().await;
-    if let Err((status, msg)) = check_admin(&state_guard, &auth_user).await {
-        return (status, Json(ApiResponse::error("FORBIDDEN", msg)));
-    }
-
-    let host = state_guard
-        .db
-        .get_setting("smtp_host")
-        .await
-        .ok()
-        .flatten()
-        .unwrap_or_default();
-    let port = state_guard
-        .db
-        .get_setting("smtp_port")
-        .await
-        .ok()
-        .flatten()
-        .unwrap_or_else(|| "587".to_string());
-    let username = state_guard
-        .db
-        .get_setting("smtp_username")
-        .await
-        .ok()
-        .flatten()
-        .unwrap_or_default();
-    let has_password = state_guard
-        .db
-        .get_setting("smtp_password")
-        .await
-        .ok()
-        .flatten()
-        .is_some_and(|p| !p.is_empty());
-    let from = state_guard
-        .db
-        .get_setting("smtp_from")
-        .await
-        .ok()
-        .flatten()
-        .unwrap_or_default();
-    let enabled = state_guard
-        .db
-        .get_setting("smtp_enabled")
-        .await
-        .ok()
-        .flatten()
-        .unwrap_or_else(|| "false".to_string());
-
-    (
-        StatusCode::OK,
-        Json(ApiResponse::success(serde_json::json!({
-            "smtp_host": host,
-            "smtp_port": port.parse::<i32>().unwrap_or(587),
-            "smtp_username": username,
-            "smtp_password": if has_password { "********" } else { "" },
-            "smtp_from": from,
-            "smtp_enabled": enabled.parse::<bool>().unwrap_or(false),
-        }))),
-    )
-}
-
-/// Request body for email settings update
-#[derive(Debug, Deserialize, utoipa::ToSchema)]
-pub struct EmailSettingsRequest {
-    #[serde(alias = "smtp_host")]
-    host: Option<String>,
-    #[serde(alias = "smtp_port")]
-    port: Option<i32>,
-    #[serde(alias = "smtp_username")]
-    username: Option<String>,
-    #[serde(alias = "smtp_password")]
-    password: Option<String>,
-    #[serde(alias = "smtp_from")]
-    from: Option<String>,
-    #[serde(alias = "smtp_enabled")]
-    enabled: Option<bool>,
-}
-
-/// `PUT /api/v1/admin/settings/email` — update SMTP settings
-#[utoipa::path(
-    put,
-    path = "/api/v1/admin/settings/email",
-    tag = "Admin",
-    summary = "Update email settings",
-    description = "Update SMTP settings for outgoing emails. Admin only.",
-    security(("bearer_auth" = []))
-)]
-pub async fn admin_update_email_settings(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-    Json(req): Json<EmailSettingsRequest>,
-) -> (StatusCode, Json<ApiResponse<serde_json::Value>>) {
-    let state_guard = state.read().await;
-    if let Err((status, msg)) = check_admin(&state_guard, &auth_user).await {
-        return (status, Json(ApiResponse::error("FORBIDDEN", msg)));
-    }
-
-    let db = &state_guard.db;
-
-    if let Some(host) = &req.host {
-        let _ = db.set_setting("smtp_host", host).await;
-    }
-    if let Some(port) = req.port {
-        let _ = db.set_setting("smtp_port", &port.to_string()).await;
-    }
-    if let Some(username) = &req.username {
-        let _ = db.set_setting("smtp_username", username).await;
-    }
-    if let Some(password) = &req.password {
-        // Don't overwrite with the masked placeholder
-        if password != "********" {
-            let _ = db.set_setting("smtp_password", password).await;
-        }
-    }
-    if let Some(from) = &req.from {
-        let _ = db.set_setting("smtp_from", from).await;
-    }
-    if let Some(enabled) = req.enabled {
-        let _ = db.set_setting("smtp_enabled", &enabled.to_string()).await;
-    }
-
-    AuditEntry::new(AuditEventType::ConfigChanged)
-        .user(auth_user.user_id, "admin")
-        .resource("settings", "email")
-        .log();
-
-    (
-        StatusCode::OK,
-        Json(ApiResponse::success(
-            serde_json::json!({"message": "Email settings updated"}),
-        )),
-    )
-}
-
-// ═══════════════════════════════════════════════════════════════════════════════
-// ADMIN: PRIVACY SETTINGS
-// ═══════════════════════════════════════════════════════════════════════════════
-
-/// `GET /api/v1/admin/privacy` — return privacy/GDPR settings
-#[utoipa::path(
-    get,
-    path = "/api/v1/admin/privacy",
-    tag = "Admin",
-    summary = "Get privacy settings",
-    description = "Return privacy and GDPR settings. Admin only.",
-    security(("bearer_auth" = []))
-)]
-pub async fn admin_get_privacy(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-) -> (StatusCode, Json<ApiResponse<serde_json::Value>>) {
-    let state_guard = state.read().await;
-    if let Err((status, msg)) = check_admin(&state_guard, &auth_user).await {
-        return (status, Json(ApiResponse::error("FORBIDDEN", msg)));
-    }
-
-    let db = &state_guard.db;
-
-    let privacy_policy_url = db
-        .get_setting("privacy_policy_url")
-        .await
-        .ok()
-        .flatten()
-        .unwrap_or_default();
-    let data_retention_days = db
-        .get_setting("data_retention_days")
-        .await
-        .ok()
-        .flatten()
-        .unwrap_or_else(|| "365".to_string());
-    let require_consent = db
-        .get_setting("require_consent")
-        .await
-        .ok()
-        .flatten()
-        .unwrap_or_else(|| "true".to_string());
-    let anonymize_on_delete = db
-        .get_setting("anonymize_on_delete")
-        .await
-        .ok()
-        .flatten()
-        .unwrap_or_else(|| "true".to_string());
-
-    (
-        StatusCode::OK,
-        Json(ApiResponse::success(serde_json::json!({
-            "privacy_policy_url": privacy_policy_url,
-            "data_retention_days": data_retention_days.parse::<i32>().unwrap_or(365),
-            "require_consent": require_consent.parse::<bool>().unwrap_or(true),
-            "anonymize_on_delete": anonymize_on_delete.parse::<bool>().unwrap_or(true),
-        }))),
-    )
-}
-
-/// Request body for privacy settings update
-#[derive(Debug, Deserialize, utoipa::ToSchema)]
-pub struct PrivacySettingsRequest {
-    privacy_policy_url: Option<String>,
-    data_retention_days: Option<i32>,
-    require_consent: Option<bool>,
-    anonymize_on_delete: Option<bool>,
-}
-
-/// `PUT /api/v1/admin/privacy` — update privacy settings
-#[utoipa::path(
-    put,
-    path = "/api/v1/admin/privacy",
-    tag = "Admin",
-    summary = "Update privacy settings",
-    description = "Update privacy and GDPR settings. Admin only.",
-    security(("bearer_auth" = []))
-)]
-pub async fn admin_update_privacy(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-    Json(req): Json<PrivacySettingsRequest>,
-) -> (StatusCode, Json<ApiResponse<serde_json::Value>>) {
-    let state_guard = state.read().await;
-    if let Err((status, msg)) = check_admin(&state_guard, &auth_user).await {
-        return (status, Json(ApiResponse::error("FORBIDDEN", msg)));
-    }
-
-    let db = &state_guard.db;
-
-    if let Some(url) = &req.privacy_policy_url {
-        let _ = db.set_setting("privacy_policy_url", url).await;
-    }
-    if let Some(days) = req.data_retention_days {
-        if days < 1 {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(ApiResponse::error(
-                    "INVALID_INPUT",
-                    "data_retention_days must be >= 1",
-                )),
-            );
-        }
-        let _ = db
-            .set_setting("data_retention_days", &days.to_string())
-            .await;
-    }
-    if let Some(consent) = req.require_consent {
-        let _ = db
-            .set_setting("require_consent", &consent.to_string())
-            .await;
-    }
-    if let Some(anonymize) = req.anonymize_on_delete {
-        let _ = db
-            .set_setting("anonymize_on_delete", &anonymize.to_string())
-            .await;
-    }
-
-    AuditEntry::new(AuditEventType::ConfigChanged)
-        .user(auth_user.user_id, "admin")
-        .resource("settings", "privacy")
-        .log();
-
-    // Return current state
-    let privacy_policy_url = db
-        .get_setting("privacy_policy_url")
-        .await
-        .ok()
-        .flatten()
-        .unwrap_or_default();
-    let data_retention_days = db
-        .get_setting("data_retention_days")
-        .await
-        .ok()
-        .flatten()
-        .unwrap_or_else(|| "365".to_string());
-    let require_consent = db
-        .get_setting("require_consent")
-        .await
-        .ok()
-        .flatten()
-        .unwrap_or_else(|| "true".to_string());
-    let anonymize_on_delete = db
-        .get_setting("anonymize_on_delete")
-        .await
-        .ok()
-        .flatten()
-        .unwrap_or_else(|| "true".to_string());
-
-    (
-        StatusCode::OK,
-        Json(ApiResponse::success(serde_json::json!({
-            "privacy_policy_url": privacy_policy_url,
-            "data_retention_days": data_retention_days.parse::<i32>().unwrap_or(365),
-            "require_consent": require_consent.parse::<bool>().unwrap_or(true),
-            "anonymize_on_delete": anonymize_on_delete.parse::<bool>().unwrap_or(true),
-        }))),
-    )
-}
-
-// ═══════════════════════════════════════════════════════════════════════════════
-// ADMIN: UPDATE USER
-// ═══════════════════════════════════════════════════════════════════════════════
-
-/// Request body for admin user update
-#[derive(Debug, Deserialize, utoipa::ToSchema)]
-pub struct AdminUpdateUserRequest {
-    name: Option<String>,
-    email: Option<String>,
-    role: Option<String>,
-    is_active: Option<bool>,
-}
-
-/// `PUT /api/v1/admin/users/{id}/update` — admin can update user details
-#[utoipa::path(
-    put,
-    path = "/api/v1/admin/users/{id}/update",
-    tag = "Admin",
-    summary = "Update user details",
-    description = "Admin can update any user's details (name, email, department, etc.).",
-    security(("bearer_auth" = []))
-)]
-pub async fn admin_update_user(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-    Path(id): Path<String>,
-    Json(req): Json<AdminUpdateUserRequest>,
-) -> (StatusCode, Json<ApiResponse<serde_json::Value>>) {
-    let state_guard = state.read().await;
-    if let Err((status, msg)) = check_admin(&state_guard, &auth_user).await {
-        return (status, Json(ApiResponse::error("FORBIDDEN", msg)));
-    }
-
-    let mut user = match state_guard.db.get_user(&id).await {
-        Ok(Some(u)) => u,
-        Ok(None) => {
-            return (
-                StatusCode::NOT_FOUND,
-                Json(ApiResponse::error("NOT_FOUND", "User not found")),
-            );
-        }
-        Err(e) => {
-            tracing::error!("Database error: {}", e);
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ApiResponse::error("SERVER_ERROR", "Internal server error")),
-            );
-        }
-    };
-
-    if let Some(name) = req.name {
-        user.name = name;
-    }
-    if let Some(email) = req.email {
-        // Basic email validation
-        if !email.contains('@') || email.len() < 5 {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(ApiResponse::error("INVALID_INPUT", "Invalid email address")),
-            );
-        }
-        user.email = email;
-    }
-    if let Some(role_str) = req.role {
-        let new_role = match role_str.to_lowercase().as_str() {
-            "user" => UserRole::User,
-            "premium" => UserRole::Premium,
-            "admin" => UserRole::Admin,
-            "superadmin" => {
-                // Only SuperAdmin can assign SuperAdmin
-                let caller = state_guard
-                    .db
-                    .get_user(&auth_user.user_id.to_string())
-                    .await
-                    .ok()
-                    .flatten();
-                if caller.map(|c| c.role) != Some(UserRole::SuperAdmin) {
-                    return (
-                        StatusCode::FORBIDDEN,
-                        Json(ApiResponse::error(
-                            "FORBIDDEN",
-                            "Only SuperAdmin can assign SuperAdmin role",
-                        )),
-                    );
-                }
-                UserRole::SuperAdmin
-            }
-            _ => {
-                return (
-                    StatusCode::BAD_REQUEST,
-                    Json(ApiResponse::error(
-                        "INVALID_INPUT",
-                        "Role must be user, premium, admin, or superadmin",
-                    )),
-                );
-            }
-        };
-        user.role = new_role;
-    }
-    if let Some(active) = req.is_active {
-        user.is_active = active;
-    }
-    user.updated_at = Utc::now();
-
-    if let Err(e) = state_guard.db.save_user(&user).await {
-        tracing::error!("Failed to update user: {}", e);
-        return (
-            StatusCode::INTERNAL_SERVER_ERROR,
-            Json(ApiResponse::error("SERVER_ERROR", "Failed to update user")),
-        );
-    }
-
-    AuditEntry::new(AuditEventType::UserUpdated)
-        .user(auth_user.user_id, "admin")
-        .resource("user", &id)
-        .log();
-
-    (
-        StatusCode::OK,
-        Json(ApiResponse::success(serde_json::json!({
-            "id": user.id.to_string(),
-            "username": user.username,
-            "email": user.email,
-            "name": user.name,
-            "role": format!("{:?}", user.role).to_lowercase(),
-            "is_active": user.is_active,
-        }))),
-    )
-}
-
-// ═══════════════════════════════════════════════════════════════════════════════
 // BOOKING UPDATE (PATCH /api/v1/bookings/{id})
 // ═══════════════════════════════════════════════════════════════════════════════
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    // ─── normalize_metric_path ─────────────────────────────────────────
-
-    #[test]
-    fn test_normalize_metric_path_uuid() {
-        let path = "/api/v1/lots/550e8400-e29b-41d4-a716-446655440000/slots";
-        assert_eq!(normalize_metric_path(path), "/api/v1/lots/:id/slots");
-    }
-
-    #[test]
-    fn test_normalize_metric_path_numeric() {
-        let path = "/api/v1/bookings/12345";
-        assert_eq!(normalize_metric_path(path), "/api/v1/bookings/:id");
-    }
-
-    #[test]
-    fn test_normalize_metric_path_no_ids() {
-        let path = "/api/v1/lots";
-        assert_eq!(normalize_metric_path(path), "/api/v1/lots");
-    }
-
-    #[test]
-    fn test_normalize_metric_path_multiple_ids() {
-        let path = "/api/v1/lots/550e8400-e29b-41d4-a716-446655440000/slots/42";
-        assert_eq!(normalize_metric_path(path), "/api/v1/lots/:id/slots/:id");
-    }
-
-    #[test]
-    fn test_normalize_metric_path_root() {
-        assert_eq!(normalize_metric_path("/health"), "/health");
-        assert_eq!(normalize_metric_path("/metrics"), "/metrics");
-    }
-
-    // ─── ImpressumData serde ───────────────────────────────────────────
-
-    #[test]
-    fn test_impressum_data_default() {
-        let data = ImpressumData::default();
-        assert_eq!(data.provider_name, "");
-        assert_eq!(data.country, "");
-    }
-
-    #[test]
-    fn test_impressum_data_roundtrip() {
-        let data = ImpressumData {
-            provider_name: "ParkCorp GmbH".to_string(),
-            provider_legal_form: "GmbH".to_string(),
-            street: "Musterstr. 1".to_string(),
-            zip_city: "12345 Berlin".to_string(),
-            country: "DE".to_string(),
-            email: "info@parkcorp.de".to_string(),
-            phone: "+49 30 123456".to_string(),
-            register_court: "Amtsgericht Berlin".to_string(),
-            register_number: "HRB 12345".to_string(),
-            vat_id: "DE123456789".to_string(),
-            responsible_person: "Max Mustermann".to_string(),
-            custom_text: "".to_string(),
-        };
-        let json = serde_json::to_string(&data).unwrap();
-        let back: ImpressumData = serde_json::from_str(&json).unwrap();
-        assert_eq!(back.provider_name, "ParkCorp GmbH");
-        assert_eq!(back.vat_id, "DE123456789");
-    }
-
-    // ─── UpdateCurrentUserRequest serde ────────────────────────────────
-
-    #[test]
-    fn test_update_current_user_request_full() {
-        let json =
-            r#"{"name":"New Name","phone":"+49123","picture":"https://img.example/pic.jpg"}"#;
-        let req: UpdateCurrentUserRequest = serde_json::from_str(json).unwrap();
-        assert_eq!(req.name.as_deref(), Some("New Name"));
-        assert_eq!(req.phone.as_deref(), Some("+49123"));
-        assert_eq!(req.picture.as_deref(), Some("https://img.example/pic.jpg"));
-    }
-
-    #[test]
-    fn test_update_current_user_request_empty() {
-        let json = r#"{}"#;
-        let req: UpdateCurrentUserRequest = serde_json::from_str(json).unwrap();
-        assert!(req.name.is_none());
-        assert!(req.phone.is_none());
-        assert!(req.picture.is_none());
-    }
-
-    // ─── UpdateUserRoleRequest / UpdateUserStatusRequest ───────────────
-
-    #[test]
-    fn test_update_user_role_request() {
-        let json = r#"{"role":"admin"}"#;
-        let req: UpdateUserRoleRequest = serde_json::from_str(json).unwrap();
-        assert_eq!(req.role, "admin");
-    }
-
-    #[test]
-    fn test_update_user_status_request() {
-        let json = r#"{"status":"active"}"#;
-        let req: UpdateUserStatusRequest = serde_json::from_str(json).unwrap();
-        assert_eq!(req.status, "active");
-    }
-
-    // ─── ChangePasswordRequest ─────────────────────────────────────────
-
-    #[test]
-    fn test_change_password_request() {
-        let json = r#"{"current_password":"old","new_password":"NewSecure123!"}"#;
-        let req: ChangePasswordRequest = serde_json::from_str(json).unwrap();
-        assert_eq!(req.current_password, "old");
-        assert_eq!(req.new_password, "NewSecure123!");
-    }
-
-    // ─── AdminResetRequest ─────────────────────────────────────────────
-
-    #[test]
-    fn test_admin_reset_request() {
-        let json = r#"{"confirm":"RESET"}"#;
-        let req: AdminResetRequest = serde_json::from_str(json).unwrap();
-        assert_eq!(req.confirm, "RESET");
-    }
-
-    // ─── AutoReleaseSettingsRequest ────────────────────────────────────
-
-    #[test]
-    fn test_auto_release_settings_request() {
-        let json = r#"{"auto_release_enabled":true,"auto_release_minutes":15}"#;
-        let req: AutoReleaseSettingsRequest = serde_json::from_str(json).unwrap();
-        assert_eq!(req.auto_release_enabled, Some(true));
-        assert_eq!(req.auto_release_minutes, Some(15));
-    }
-
-    #[test]
-    fn test_auto_release_settings_request_partial() {
-        let json = r#"{"auto_release_minutes":45}"#;
-        let req: AutoReleaseSettingsRequest = serde_json::from_str(json).unwrap();
-        assert!(req.auto_release_enabled.is_none());
-        assert_eq!(req.auto_release_minutes, Some(45));
-    }
-
-    // ─── EmailSettingsRequest ──────────────────────────────────────────
-
-    #[test]
-    fn test_email_settings_request_full() {
-        let json = r#"{
-            "smtp_host":"smtp.example.com",
-            "smtp_port":587,
-            "smtp_username":"user@example.com",
-            "smtp_password":"secret",
-            "smtp_from":"noreply@example.com",
-            "smtp_enabled":true
-        }"#;
-        let req: EmailSettingsRequest = serde_json::from_str(json).unwrap();
-        assert_eq!(req.host.as_deref(), Some("smtp.example.com"));
-        assert_eq!(req.port, Some(587));
-        assert_eq!(req.enabled, Some(true));
-    }
-
-    #[test]
-    fn test_email_settings_request_empty() {
-        let json = r#"{}"#;
-        let req: EmailSettingsRequest = serde_json::from_str(json).unwrap();
-        assert!(req.host.is_none());
-        assert!(req.port.is_none());
-        assert!(req.enabled.is_none());
-    }
-
-    // ─── PrivacySettingsRequest ────────────────────────────────────────
-
-    #[test]
-    fn test_privacy_settings_request() {
-        let json = r#"{
-            "privacy_policy_url":"https://example.com/privacy",
-            "data_retention_days":365,
-            "require_consent":true,
-            "anonymize_on_delete":true
-        }"#;
-        let req: PrivacySettingsRequest = serde_json::from_str(json).unwrap();
-        assert_eq!(
-            req.privacy_policy_url.as_deref(),
-            Some("https://example.com/privacy")
-        );
-        assert_eq!(req.data_retention_days, Some(365));
-        assert_eq!(req.require_consent, Some(true));
-        assert_eq!(req.anonymize_on_delete, Some(true));
-    }
-
-    // ─── AdminUpdateUserRequest ────────────────────────────────────────
-
-    #[test]
-    fn test_admin_update_user_request_full() {
-        let json =
-            r#"{"name":"Updated","email":"new@example.com","role":"admin","is_active":false}"#;
-        let req: AdminUpdateUserRequest = serde_json::from_str(json).unwrap();
-        assert_eq!(req.name.as_deref(), Some("Updated"));
-        assert_eq!(req.email.as_deref(), Some("new@example.com"));
-        assert_eq!(req.role.as_deref(), Some("admin"));
-        assert_eq!(req.is_active, Some(false));
-    }
-
-    #[test]
-    fn test_admin_update_user_request_partial() {
-        let json = r#"{"is_active":true}"#;
-        let req: AdminUpdateUserRequest = serde_json::from_str(json).unwrap();
-        assert!(req.name.is_none());
-        assert!(req.email.is_none());
-        assert!(req.role.is_none());
-        assert_eq!(req.is_active, Some(true));
-    }
-
-    // ─── UpdatePreferencesRequest ──────────────────────────────────────
-
-    #[test]
-    fn test_update_preferences_request() {
-        let json = r#"{"language":"de","theme":"dark","notifications_enabled":false}"#;
-        let req: UpdatePreferencesRequest = serde_json::from_str(json).unwrap();
-        assert_eq!(req.language.as_deref(), Some("de"));
-        assert_eq!(req.theme.as_deref(), Some("dark"));
-        assert_eq!(req.notifications_enabled, Some(false));
-        assert!(req.email_reminders.is_none());
-        assert!(req.default_duration_minutes.is_none());
-    }
-}
