+++ parkhub-server/src/api/system.rs
@@ -1,0 +1,347 @@
+//! System handlers: health checks, version, maintenance, handshake, server status, metrics.
+//!
+//! These endpoints provide operational observability and protocol negotiation.
+//! None require authentication (they are public/infrastructure endpoints).
+
+use axum::{
+    body::Body,
+    extract::State,
+    http::{HeaderName, HeaderValue, Request, StatusCode},
+    middleware::Next,
+    response::{IntoResponse, Response},
+    Json,
+};
+use std::sync::Arc;
+use tokio::sync::RwLock;
+use tracing::Instrument;
+
+use parkhub_common::{
+    ApiResponse, HandshakeRequest, HandshakeResponse, ServerStatus, PROTOCOL_VERSION,
+};
+
+use crate::AppState;
+
+type SharedState = Arc<RwLock<AppState>>;
+
+// ═══════════════════════════════════════════════════════════════════════════════
+// HEALTH & DISCOVERY
+// ═══════════════════════════════════════════════════════════════════════════════
+
+#[utoipa::path(
+    get,
+    path = "/health",
+    tag = "Health",
+    summary = "Basic health check",
+    description = "Returns 200 OK when the server is running.",
+    responses((status = 200, description = "Healthy"))
+)]
+#[tracing::instrument]
+pub async fn health_check() -> &'static str {
+    "OK"
+}
+
+/// Kubernetes liveness probe - just checks if the service is running
+#[utoipa::path(
+    get,
+    path = "/health/live",
+    tag = "Health",
+    summary = "Kubernetes liveness probe",
+    description = "Returns 200 if the process is alive.",
+    responses((status = 200, description = "Alive"))
+)]
+#[tracing::instrument]
+pub async fn liveness_check() -> StatusCode {
+    StatusCode::OK
+}
+
+/// Kubernetes readiness probe - checks if the service can handle traffic.
+///
+/// Returns only a boolean `ready` field. Internal error details are logged
+/// server-side but never exposed in the response body.
+#[utoipa::path(
+    get,
+    path = "/health/ready",
+    tag = "Health",
+    summary = "Kubernetes readiness probe",
+    description = "Returns 200 when the service can accept traffic.",
+    responses(
+        (status = 200, description = "Ready"),
+        (status = 503, description = "Not ready")
+    )
+)]
+#[tracing::instrument(skip(state))]
+pub async fn readiness_check(State(state): State<SharedState>) -> impl IntoResponse {
+    let state = state.read().await;
+    match state.db.stats().await {
+        Ok(_) => (StatusCode::OK, Json(serde_json::json!({"ready": true}))),
+        Err(e) => {
+            tracing::error!(error = %e, "Readiness check failed — database unavailable");
+            (
+                StatusCode::SERVICE_UNAVAILABLE,
+                Json(serde_json::json!({"ready": false})),
+            )
+        }
+    }
+}
+
+/// `GET /api/v1/system/version` — server version information
+pub async fn system_version() -> Json<serde_json::Value> {
+    Json(serde_json::json!({
+        "version": env!("CARGO_PKG_VERSION"),
+        "name": env!("CARGO_PKG_NAME"),
+    }))
+}
+
+/// `GET /api/v1/system/maintenance` — maintenance mode status
+pub async fn system_maintenance(State(state): State<SharedState>) -> Json<serde_json::Value> {
+    let state = state.read().await;
+    let maintenance = match state.db.get_setting("maintenance_mode").await {
+        Ok(Some(v)) => v == "true",
+        _ => false,
+    };
+    Json(serde_json::json!({
+        "maintenance_mode": maintenance,
+        "message": if maintenance { "System is under maintenance" } else { "" }
+    }))
+}
+
+#[utoipa::path(
+    post,
+    path = "/handshake",
+    tag = "Health",
+    summary = "Protocol handshake",
+    description = "Verifies protocol version compatibility between client and server.",
+    responses((status = 200, description = "Handshake result"))
+)]
+pub async fn handshake(
+    State(state): State<SharedState>,
+    Json(request): Json<HandshakeRequest>,
+) -> Json<ApiResponse<HandshakeResponse>> {
+    let state = state.read().await;
+
+    // Check protocol version compatibility
+    if request.protocol_version != PROTOCOL_VERSION {
+        return Json(ApiResponse::error(
+            "PROTOCOL_MISMATCH",
+            format!(
+                "Protocol version mismatch: server={}, client={}",
+                PROTOCOL_VERSION, request.protocol_version
+            ),
+        ));
+    }
+
+    Json(ApiResponse::success(HandshakeResponse {
+        server_name: state.config.server_name.clone(),
+        server_version: env!("CARGO_PKG_VERSION").to_string(),
+        protocol_version: PROTOCOL_VERSION.to_string(),
+        requires_auth: true,
+        certificate_fingerprint: String::new(),
+    }))
+}
+
+#[utoipa::path(
+    get,
+    path = "/status",
+    tag = "Health",
+    summary = "Server status overview",
+    description = "Returns aggregate server statistics.",
+    responses((status = 200, description = "Server status"))
+)]
+pub async fn server_status(State(state): State<SharedState>) -> Json<ApiResponse<ServerStatus>> {
+    let db_stats = {
+        let state = state.read().await;
+        state.db.stats().await.unwrap_or(crate::db::DatabaseStats {
+            users: 0,
+            bookings: 0,
+            parking_lots: 0,
+            slots: 0,
+            sessions: 0,
+            vehicles: 0,
+        })
+    };
+
+    Json(ApiResponse::success(ServerStatus {
+        uptime_seconds: 0,
+        connected_clients: 0,
+        total_users: u32::try_from(db_stats.users).unwrap_or(u32::MAX),
+        total_bookings: u32::try_from(db_stats.bookings).unwrap_or(u32::MAX),
+        database_size_bytes: 0,
+    }))
+}
+
+// ═══════════════════════════════════════════════════════════════════════════════
+// MIDDLEWARE
+// ═══════════════════════════════════════════════════════════════════════════════
+
+/// Middleware that adds security-related response headers to every request.
+///
+/// - `X-Content-Type-Options`: prevents MIME sniffing
+/// - `X-Frame-Options`: prevents clickjacking
+/// - `Content-Security-Policy`: restricts resource origins
+/// - `Referrer-Policy`: limits referrer leakage
+/// - `Permissions-Policy`: disables unneeded browser features
+/// - `Strict-Transport-Security`: enforces HTTPS for 1 year including subdomains
+pub async fn security_headers_middleware(request: Request<Body>, next: Next) -> Response {
+    let mut response = next.run(request).await;
+    let headers = response.headers_mut();
+
+    // Safety: all header names and values below are valid ASCII strings.
+    headers.insert(
+        HeaderName::from_static("x-content-type-options"),
+        HeaderValue::from_static("nosniff"),
+    );
+    headers.insert(
+        HeaderName::from_static("x-frame-options"),
+        HeaderValue::from_static("DENY"),
+    );
+    headers.insert(
+        HeaderName::from_static("content-security-policy"),
+        HeaderValue::from_static(
+            "default-src 'self'; \
+             script-src 'self'; \
+             style-src 'self'; \
+             img-src 'self' data:; \
+             font-src 'self'; \
+             connect-src 'self' ws: wss:; \
+             frame-ancestors 'none'",
+        ),
+    );
+    headers.insert(
+        HeaderName::from_static("referrer-policy"),
+        HeaderValue::from_static("strict-origin-when-cross-origin"),
+    );
+    headers.insert(
+        HeaderName::from_static("permissions-policy"),
+        HeaderValue::from_static("geolocation=(), camera=(), microphone=()"),
+    );
+    headers.insert(
+        HeaderName::from_static("strict-transport-security"),
+        HeaderValue::from_static("max-age=31536000; includeSubDomains; preload"),
+    );
+
+    response
+}
+
+// ═══════════════════════════════════════════════════════════════════════════════
+// REQUEST ID TRACING MIDDLEWARE
+// ═══════════════════════════════════════════════════════════════════════════════
+
+/// Attaches the `x-request-id` header value (set by `SetRequestIdLayer`) to the
+/// current tracing span so that every log line within this request includes the
+/// request ID.  Also ensures the header is copied to the response.
+pub async fn request_id_tracing_middleware(request: Request<Body>, next: Next) -> Response {
+    let request_id = request
+        .headers()
+        .get("x-request-id")
+        .and_then(|v| v.to_str().ok())
+        .unwrap_or("-")
+        .to_owned();
+
+    let span = tracing::info_span!("request", request_id = %request_id);
+    let guard = span.enter();
+
+    let mut response = {
+        // Drop the span guard before awaiting so the span covers the full
+        // request lifecycle via the instrument approach.
+        drop(guard);
+        next.run(request).instrument(span).await
+    };
+
+    // Ensure x-request-id is on the response (belt-and-suspenders with PropagateRequestIdLayer)
+    if let Ok(val) = HeaderValue::from_str(&request_id) {
+        response
+            .headers_mut()
+            .entry(HeaderName::from_static("x-request-id"))
+            .or_insert(val);
+    }
+
+    response
+}
+
+// ═══════════════════════════════════════════════════════════════════════════════
+// HTTP METRICS MIDDLEWARE
+// ═══════════════════════════════════════════════════════════════════════════════
+
+/// Records HTTP request metrics (method, path, status, duration) for Prometheus
+/// and emits a structured log line for every request.
+pub async fn http_metrics_middleware(request: Request<Body>, next: Next) -> Response {
+    let method = request.method().to_string();
+    let path = request.uri().path().to_string();
+    let start = std::time::Instant::now();
+
+    let response = next.run(request).await;
+
+    let status = response.status().as_u16();
+    let duration = start.elapsed();
+
+    // Normalize path to avoid high-cardinality labels (strip UUIDs/IDs)
+    let normalized = normalize_metric_path(&path);
+    crate::metrics::record_http_request(&method, &normalized, status, duration);
+
+    // Structured request log — every request gets one line with key fields
+    tracing::info!(
+        http.method = %method,
+        http.path = %path,
+        http.status = status,
+        http.latency_ms = u64::try_from(duration.as_millis()).unwrap_or(u64::MAX),
+        "request completed"
+    );
+
+    response
+}
+
+/// Collapse dynamic path segments (UUIDs, numeric IDs) into placeholders
+/// to keep Prometheus label cardinality bounded.
+pub fn normalize_metric_path(path: &str) -> String {
+    path.split('/')
+        .map(|s| {
+            let is_uuid = s.len() == 36 && s.chars().filter(|c| *c == '-').count() == 4;
+            let is_numeric = !s.is_empty() && s.chars().all(|c| c.is_ascii_digit());
+            if is_uuid || is_numeric {
+                ":id"
+            } else {
+                s
+            }
+        })
+        .collect::<Vec<_>>()
+        .join("/")
+}
+
+// ═══════════════════════════════════════════════════════════════════════════════
+// Tests
+// ═══════════════════════════════════════════════════════════════════════════════
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_normalize_metric_path_uuid() {
+        let path = "/api/v1/lots/550e8400-e29b-41d4-a716-446655440000/slots";
+        assert_eq!(normalize_metric_path(path), "/api/v1/lots/:id/slots");
+    }
+
+    #[test]
+    fn test_normalize_metric_path_numeric() {
+        let path = "/api/v1/bookings/12345";
+        assert_eq!(normalize_metric_path(path), "/api/v1/bookings/:id");
+    }
+
+    #[test]
+    fn test_normalize_metric_path_no_ids() {
+        let path = "/api/v1/lots";
+        assert_eq!(normalize_metric_path(path), "/api/v1/lots");
+    }
+
+    #[test]
+    fn test_normalize_metric_path_multiple_ids() {
+        let path = "/api/v1/lots/550e8400-e29b-41d4-a716-446655440000/slots/42";
+        assert_eq!(normalize_metric_path(path), "/api/v1/lots/:id/slots/:id");
+    }
+
+    #[test]
+    fn test_normalize_metric_path_root() {
+        assert_eq!(normalize_metric_path("/health"), "/health");
+        assert_eq!(normalize_metric_path("/metrics"), "/metrics");
+    }
+}
